Release 1.4 lifecycle cost collection
