+++ .claude-plugin/marketplace.json
@@ -9,10 +9,10 @@
       "source": {
         "source": "github",
         "repo": "avksp/agent-lifecycle-kit",
-        "ref": "v1.3.0"
+        "ref": "v1.4.0"
       },
       "description": "Reviewed SDD planning, budgeted execution, adapter conformance, implementation audit, and final proof.",
-      "version": "1.3.0",
+      "version": "1.4.0",
       "author": {
         "name": "Agent Lifecycle Kit contributors"
       },
