+++ docs/guides/README.ru.md
@@ -20,6 +20,7 @@ CLI и плагинов.
 - Компактный контекст и короткие снимки цели для небольших локальных моделей.
 - Учет расхода по категориям: реализация, проверка продукта, контроль цикла и
   координация.
+- Детерминированные отчеты о расходе по явным артефактам жизненного цикла.
 - Диагностика готовности без записи файлов и без запуска живых модельных
   вызовов по умолчанию.
 
@@ -53,6 +54,7 @@ agent-lifecycle adapter validate --descriptor adapters/codex/adapter.descriptor.
 - `agent-lifecycle plan`: проверки плана, замки, снимки и передача ревьюеру.
 - `agent-lifecycle workflow`: квитанции задач и финальное доказательство.
 - `agent-lifecycle audit`: проверка плана и реализации.
+- `agent-lifecycle metrics`: отчеты о расходе и их проверка.
 - `agent-lifecycle diagnostics`: редактированные пакеты доказательств.
 - `agent-lifecycle diagnose`: готовность исходного дерева без записи и живых
   вызовов.
