+++ release/notes/v1.4.0.md
@@ -1,0 +1,38 @@
+# Agent Lifecycle Kit v1.4.0 Notes
+
+Status: source release.
+
+This release turns lifecycle cost accounting into a generated artifact while
+keeping existing manual cost reports valid.
+
+## Changes
+
+- Added deterministic lifecycle cost report generation from explicit JSON
+  artifact paths.
+- Added source artifact identity, lineage binding, usage confidence counts and
+  compact cost summaries.
+- Added `agent-lifecycle metrics cost-report` with write-once report output,
+  optional compact summary output and immediate validation.
+- Split built-in schema definitions into smaller domain modules without
+  changing the public schema registry.
+- Split safe host inspection probes by adapter while preserving the public
+  inspection API.
+- Updated package and plugin metadata to `1.4.0`.
+
+## Validation
+
+- Generated report fixtures cover attested, estimated and missing usage states.
+- CLI smoke writes a report and compact summary, then checks the summary with
+  the small-context profile.
+- Manual `agent-lifecycle-cost-report.v1` inputs remain accepted by
+  `metrics cost-check`.
+- Contract tests cover the split schema registry through the existing public
+  `schema list/show` surface.
+- Adapter inspection tests cover the split host probe modules through the
+  existing inspect command surface.
+
+## Claims
+
+- Adapter maturity changes: none.
+- Production promotion claimed: false.
+- Public directory or marketplace approval claimed: false.
