+++ src/agent_lifecycle/_version.py
@@ -1,3 +1,3 @@
 """Package version."""
 
-__version__ = "1.3.0"
+__version__ = "1.4.0"
