+++ src/agent_lifecycle/contracts/context_model_schemas.py
@@ -1,0 +1,246 @@
+"""Built-in JSON schema definitions for a lifecycle contract domain."""
+
+from __future__ import annotations
+
+from typing import Any
+
+from agent_lifecycle.contracts.schema_builders import open_object_schema as _open_object_schema
+
+CONTEXT_MODEL_SCHEMAS: dict[str, dict[str, Any]] = {
+    "agent-small-context-profile.v1": {
+        "$schema": "https://json-schema.org/draft/2020-12/schema",
+        "$id": "agent-small-context-profile.v1",
+        "type": "object",
+        "additionalProperties": True,
+        "required": ["schemaVersion", "profileId", "defaultWindow", "windows", "requiredSummaryFields"],
+        "properties": {
+            "schemaVersion": {"const": "agent-small-context-profile.v1"},
+            "profileId": {"type": "string", "minLength": 1},
+            "defaultWindow": {"enum": ["4k-strict", "8k", "16k", "32k", "64k"]},
+            "windows": {"type": "object"},
+            "requiredSummaryFields": {"type": "array", "items": {"type": "string"}},
+        },
+    },
+    "agent-context-render-receipt.v1": {
+        "$schema": "https://json-schema.org/draft/2020-12/schema",
+        "$id": "agent-context-render-receipt.v1",
+        "type": "object",
+        "additionalProperties": True,
+        "required": ["schemaVersion", "status", "profileId", "window", "estimatedTokens", "checks", "envelopeDigest"],
+        "properties": {
+            "schemaVersion": {"const": "agent-context-render-receipt.v1"},
+            "status": {"enum": ["PASS", "FAIL"]},
+            "profileId": {"type": "string", "minLength": 1},
+            "window": {"enum": ["4k-strict", "8k", "16k", "32k", "64k"]},
+            "estimatedTokens": {"type": "object"},
+            "checks": {"type": "array", "items": {"type": "object"}},
+            "envelopeDigest": {"type": "string", "minLength": 64, "maxLength": 64},
+        },
+    },
+    "agent-sdd-tier-resolution.v1": {
+        "$schema": "https://json-schema.org/draft/2020-12/schema",
+        "$id": "agent-sdd-tier-resolution.v1",
+        "type": "object",
+        "additionalProperties": False,
+        "required": ["schemaVersion", "tier", "reasons", "requestDigest", "rules"],
+        "properties": {
+            "schemaVersion": {"const": "agent-sdd-tier-resolution.v1"},
+            "tier": {"enum": ["S0", "S1", "S2"]},
+            "reasons": {"type": "array", "items": {"type": "string"}},
+            "requestDigest": {"type": "string", "minLength": 64, "maxLength": 64},
+            "rules": {"type": "object"},
+        },
+    },
+    "agent-lifecycle-model-routing-profile.v1": {
+        "$schema": "https://json-schema.org/draft/2020-12/schema",
+        "$id": "agent-lifecycle-model-routing-profile.v1",
+        "type": "object",
+        "additionalProperties": True,
+        "required": ["schemaVersion", "profileId", "classes", "phaseRules"],
+        "properties": {
+            "schemaVersion": {"const": "agent-lifecycle-model-routing-profile.v1"},
+            "profileId": {"type": "string", "minLength": 1},
+            "classes": {"type": "array", "items": {"type": "string", "minLength": 1}},
+            "phaseRules": {"type": "object"},
+        },
+    },
+    "agent-lifecycle-host-model-profile.v1": {
+        "$schema": "https://json-schema.org/draft/2020-12/schema",
+        "$id": "agent-lifecycle-host-model-profile.v1",
+        "type": "object",
+        "additionalProperties": True,
+        "required": ["schemaVersion", "host", "profileId", "bindings"],
+        "properties": {
+            "schemaVersion": {"const": "agent-lifecycle-host-model-profile.v1"},
+            "host": {"type": "string", "minLength": 1},
+            "profileId": {"type": "string", "minLength": 1},
+            "bindings": {"type": "object"},
+        },
+    },
+    "agent-host-model-selection-profile.v1": {
+        "$schema": "https://json-schema.org/draft/2020-12/schema",
+        "$id": "agent-host-model-selection-profile.v1",
+        "type": "object",
+        "additionalProperties": True,
+        "required": [
+            "schemaVersion",
+            "host",
+            "profileId",
+            "budgetModeDefault",
+            "bindings",
+            "fallbackPolicies",
+            "redactionPolicy",
+        ],
+        "properties": {
+            "schemaVersion": {"const": "agent-host-model-selection-profile.v1"},
+            "host": {"type": "string", "minLength": 1},
+            "profileId": {"type": "string", "minLength": 1},
+            "budgetModeDefault": {"enum": ["metered", "subscription", "local"]},
+            "bindings": {"type": "object"},
+            "fallbackPolicies": {"type": "object"},
+            "redactionPolicy": {"type": "object"},
+        },
+    },
+    "agent-host-model-selection-receipt.v1": _open_object_schema(
+        "agent-host-model-selection-receipt.v1",
+        required=[
+            "schemaVersion",
+            "host",
+            "routeDecisionDigest",
+            "modelClass",
+            "profileDigest",
+            "bindingId",
+            "bindingDigest",
+            "providerModelHash",
+            "fallbackUsed",
+        ],
+        properties={
+            "host": {"type": "string", "minLength": 1},
+            "routeDecisionDigest": {"type": "string", "minLength": 64, "maxLength": 64},
+            "modelClass": {"type": "string", "minLength": 1},
+            "profileDigest": {"type": "string", "minLength": 64, "maxLength": 64},
+            "bindingId": {"type": "string", "minLength": 1},
+            "bindingDigest": {"type": "string", "minLength": 64, "maxLength": 64},
+            "providerModelHash": {"type": "string", "minLength": 64, "maxLength": 64},
+            "fallbackUsed": {"type": "boolean"},
+            "fallbackReason": {"type": ["string", "null"]},
+        },
+    ),
+    "agent-lifecycle-budget-exceeded-policy.v1": _open_object_schema(
+        "agent-lifecycle-budget-exceeded-policy.v1",
+        required=[
+            "schemaVersion",
+            "mode",
+            "allowedActions",
+            "forbidDowngradeForCriticalReview",
+            "maxAutoReroutesPerTask",
+            "budgetModes",
+        ],
+        properties={
+            "mode": {"enum": ["manual", "auto"]},
+            "allowedActions": {
+                "type": "array",
+                "items": {
+                    "enum": [
+                        "continue-same-route",
+                        "reroute-cheaper",
+                        "reroute-stronger",
+                        "split-task",
+                        "abort",
+                    ]
+                },
+            },
+            "forbidDowngradeForCriticalReview": {"type": "boolean"},
+            "maxAutoReroutesPerTask": {"type": "integer", "minimum": 0},
+            "defaultAutoAction": {"type": "string"},
+            "budgetModes": {"type": "object"},
+        },
+    ),
+    "agent-lifecycle-budget-decision-receipt.v1": _open_object_schema(
+        "agent-lifecycle-budget-decision-receipt.v1",
+        required=[
+            "schemaVersion",
+            "runId",
+            "taskId",
+            "attemptId",
+            "sourceRevision",
+            "priorRouteDecisionDigest",
+            "usageReceiptDigest",
+            "overrunReason",
+            "decisionMode",
+            "selectedAction",
+            "resumePhase",
+            "expectedWorkflowRevision",
+        ],
+        properties={
+            "runId": {"type": "string", "minLength": 1},
+            "taskId": {"type": "string", "minLength": 1},
+            "attemptId": {"type": "string", "minLength": 1},
+            "sourceRevision": {"type": "string", "minLength": 1},
+            "priorRouteDecisionDigest": {"type": "string", "minLength": 64, "maxLength": 64},
+            "usageReceiptDigest": {"type": "string", "minLength": 64, "maxLength": 64},
+            "overrunReason": {"type": "string", "minLength": 1},
+            "decisionMode": {"enum": ["manual", "auto"]},
+            "selectedAction": {"type": "string", "minLength": 1},
+            "operatorIdentityHash": {"type": "string"},
+            "policyDigest": {"type": "string"},
+            "nextRouteDecisionDigest": {"type": ["string", "null"]},
+            "capDeltas": {"type": "object"},
+            "resumePhase": {"type": "string", "minLength": 1},
+            "expectedWorkflowRevision": {"type": "integer", "minimum": 0},
+        },
+    ),
+    "agent-cursor-compat-evidence.v1": _open_object_schema(
+        "agent-cursor-compat-evidence.v1",
+        required=["schemaVersion", "status", "source", "checks", "productionPromotionClaimed"],
+        properties={
+            "status": {"enum": ["PASS", "FAIL"]},
+            "source": {"type": "object"},
+            "checks": {"type": "array", "items": {"type": "object"}},
+            "providerModelHashes": {"type": "object"},
+            "productionPromotionClaimed": {"const": False},
+        },
+    ),
+    "agent-lifecycle-model-route-request.v1": {
+        "$schema": "https://json-schema.org/draft/2020-12/schema",
+        "$id": "agent-lifecycle-model-route-request.v1",
+        "type": "object",
+        "additionalProperties": True,
+        "required": ["schemaVersion", "operationId", "phase", "sddTier"],
+        "properties": {
+            "schemaVersion": {"const": "agent-lifecycle-model-route-request.v1"},
+            "operationId": {"type": "string", "minLength": 1},
+            "phase": {"type": "string", "minLength": 1},
+            "sddTier": {"enum": ["S0", "S1", "S2"]},
+        },
+    },
+    "agent-lifecycle-model-route-decision.v1": {
+        "$schema": "https://json-schema.org/draft/2020-12/schema",
+        "$id": "agent-lifecycle-model-route-decision.v1",
+        "type": "object",
+        "additionalProperties": True,
+        "required": ["schemaVersion", "operationId", "modelClass", "decisionDigest"],
+        "properties": {
+            "schemaVersion": {"const": "agent-lifecycle-model-route-decision.v1"},
+            "operationId": {"type": "string", "minLength": 1},
+            "modelClass": {"type": "string", "minLength": 1},
+            "decisionDigest": {"type": "string", "minLength": 64, "maxLength": 64},
+        },
+    },
+    "agent-lifecycle-model-usage-receipt.v1": {
+        "$schema": "https://json-schema.org/draft/2020-12/schema",
+        "$id": "agent-lifecycle-model-usage-receipt.v1",
+        "type": "object",
+        "additionalProperties": True,
+        "required": ["schemaVersion", "operationId", "host", "modelClass", "providerModelHash", "usage", "attestation"],
+        "properties": {
+            "schemaVersion": {"const": "agent-lifecycle-model-usage-receipt.v1"},
+            "operationId": {"type": "string", "minLength": 1},
+            "host": {"type": "string", "minLength": 1},
+            "modelClass": {"type": "string", "minLength": 1},
+            "providerModelHash": {"type": "string", "minLength": 1},
+            "usage": {"type": "object"},
+            "attestation": {"type": "object"},
+        },
+    },
+}
