+++ src/agent_lifecycle/contracts/evidence_import_schemas.py
@@ -1,0 +1,150 @@
+"""Built-in JSON schema definitions for a lifecycle contract domain."""
+
+from __future__ import annotations
+
+from typing import Any
+
+from agent_lifecycle.contracts.schema_builders import open_object_schema as _open_object_schema
+
+EVIDENCE_IMPORT_SCHEMAS: dict[str, dict[str, Any]] = {
+    "agent-evidence-index.v1": _open_object_schema(
+        "agent-evidence-index.v1",
+        required=[
+            "schemaVersion",
+            "status",
+            "sourceOfTruth",
+            "rebuildable",
+            "enabledByDefault",
+            "activationMode",
+            "resourceCaps",
+            "artifactCount",
+            "entries",
+            "blockers",
+            "indexDigest",
+        ],
+        properties={
+            "status": {"enum": ["PASS", "FAIL"]},
+            "sourceOfTruth": {"const": False},
+            "rebuildable": {"const": True},
+            "enabledByDefault": {"const": False},
+            "activationMode": {"const": "explicit-command"},
+            "resourceCaps": {"type": "object"},
+            "artifactCount": {"type": "integer", "minimum": 0},
+            "entries": {"type": "array", "items": {"type": "object"}},
+            "blockers": {"type": "array", "items": {"type": "object"}},
+            "indexDigest": {"type": "string", "minLength": 64, "maxLength": 64},
+        },
+    ),
+    "agent-evidence-index-validation.v1": _open_object_schema(
+        "agent-evidence-index-validation.v1",
+        required=["schemaVersion", "status", "artifactCount", "blockers", "indexDigest", "validationDigest"],
+        properties={
+            "status": {"enum": ["PASS", "FAIL"]},
+            "artifactCount": {"type": "integer", "minimum": 0},
+            "blockers": {"type": "array", "items": {"type": "object"}},
+            "indexDigest": {"type": ["string", "null"], "minLength": 64, "maxLength": 64},
+            "validationDigest": {"type": "string", "minLength": 64, "maxLength": 64},
+        },
+    ),
+    "agent-evidence-search-summary.v1": _open_object_schema(
+        "agent-evidence-search-summary.v1",
+        required=["schemaVersion", "status", "sourceOfTruth", "query", "resultCount", "results", "indexDigest", "summaryDigest"],
+        properties={
+            "status": {"enum": ["PASS", "FAIL"]},
+            "sourceOfTruth": {"const": False},
+            "query": {"type": "string"},
+            "resultCount": {"type": "integer", "minimum": 0},
+            "results": {"type": "array", "items": {"type": "object"}},
+            "indexDigest": {"type": ["string", "null"], "minLength": 64, "maxLength": 64},
+            "summaryDigest": {"type": "string", "minLength": 64, "maxLength": 64},
+        },
+    ),
+    "agent-planning-import-result.v1": _open_object_schema(
+        "agent-planning-import-result.v1",
+        required=[
+            "schemaVersion",
+            "status",
+            "enabledByDefault",
+            "activationMode",
+            "sourceTrusted",
+            "candidateLifecycleStatus",
+            "reviewGates",
+            "resourceCaps",
+            "source",
+            "requiresReview",
+            "auditRequired",
+            "freezeBlocked",
+            "blockers",
+            "importDigest",
+        ],
+        properties={
+            "status": {"enum": ["PASS", "FAIL"]},
+            "enabledByDefault": {"const": False},
+            "activationMode": {"const": "explicit-command"},
+            "sourceTrusted": {"const": False},
+            "candidateLifecycleStatus": {"enum": ["DRAFT_REQUIRES_REVIEW", "BLOCKED"]},
+            "reviewGates": {"type": "array", "items": {"type": "string"}},
+            "resourceCaps": {"type": "object"},
+            "source": {"type": "object"},
+            "candidatePlan": {"type": ["object", "null"]},
+            "requiresReview": {"const": True},
+            "auditRequired": {"const": True},
+            "freezeBlocked": {"const": True},
+            "blockers": {"type": "array", "items": {"type": "object"}},
+            "importDigest": {"type": "string", "minLength": 64, "maxLength": 64},
+        },
+    ),
+    "agent-planning-import-validation.v1": _open_object_schema(
+        "agent-planning-import-validation.v1",
+        required=["schemaVersion", "status", "freezeBlocked", "auditRequired", "requiresReview", "blockers", "validationDigest"],
+        properties={
+            "status": {"enum": ["PASS", "FAIL"]},
+            "candidateDigest": {"type": ["string", "null"], "minLength": 64, "maxLength": 64},
+            "freezeBlocked": {"type": "boolean"},
+            "auditRequired": {"type": "boolean"},
+            "requiresReview": {"type": "boolean"},
+            "blockers": {"type": "array", "items": {"type": "object"}},
+            "validationDigest": {"type": "string", "minLength": 64, "maxLength": 64},
+        },
+    ),
+    "agent-skill-improvement-proposal.v1": _open_object_schema(
+        "agent-skill-improvement-proposal.v1",
+        required=[
+            "schemaVersion",
+            "proposalId",
+            "affectedSkill",
+            "status",
+            "rationale",
+            "expectedBehavior",
+            "requiredTests",
+            "requiresReview",
+            "autoApply",
+            "applied",
+        ],
+        properties={
+            "proposalId": {"type": "string", "minLength": 1},
+            "affectedSkill": {"type": "string", "minLength": 1},
+            "status": {"const": "PROPOSED"},
+            "rationale": {"type": "string", "minLength": 1},
+            "expectedBehavior": {"type": "string", "minLength": 1},
+            "requiredTests": {"type": "array", "items": {"type": "string", "minLength": 1}},
+            "requiresReview": {"const": True},
+            "autoApply": {"const": False},
+            "applied": {"const": False},
+        },
+    ),
+    "agent-skill-improvement-proposal-validation.v1": _open_object_schema(
+        "agent-skill-improvement-proposal-validation.v1",
+        required=["schemaVersion", "status", "proposalId", "requiresReview", "autoApply", "applied", "blockers", "proposalDigest", "validationDigest"],
+        properties={
+            "status": {"enum": ["PASS", "FAIL"]},
+            "proposalId": {"type": ["string", "null"]},
+            "requiresReview": {"type": "boolean"},
+            "autoApply": {"type": ["boolean", "null"]},
+            "applied": {"type": ["boolean", "null"]},
+            "blockers": {"type": "array", "items": {"type": "object"}},
+            "proposalDigest": {"type": "string", "minLength": 64, "maxLength": 64},
+            "validationDigest": {"type": "string", "minLength": 64, "maxLength": 64},
+        },
+    ),
+}
