+++ src/agent_lifecycle/contracts/runner_worktree_schemas.py
@@ -1,0 +1,230 @@
+"""Built-in JSON schema definitions for a lifecycle contract domain."""
+
+from __future__ import annotations
+
+from typing import Any
+
+from agent_lifecycle.contracts.schema_builders import open_object_schema as _open_object_schema
+
+RUNNER_WORKTREE_SCHEMAS: dict[str, dict[str, Any]] = {
+    "agent-worktree-isolation-policy.v1": _open_object_schema(
+        "agent-worktree-isolation-policy.v1",
+        required=["schemaVersion", "worktreeRoot", "preserveFailedAttempts", "cleanupRequiresOperator"],
+        properties={
+            "worktreeRoot": {"type": "string", "minLength": 1},
+            "allowedWriteRoots": {"type": "array", "items": {"type": "string", "minLength": 1}},
+            "preserveFailedAttempts": {"type": "boolean"},
+            "cleanupRequiresOperator": {"type": "boolean"},
+        },
+    ),
+    "agent-worktree-isolation-policy-validation.v1": _open_object_schema(
+        "agent-worktree-isolation-policy-validation.v1",
+        required=["schemaVersion", "status", "worktreeRoot", "allowedWriteRoots", "preserveFailedAttempts", "cleanupRequiresOperator", "policyDigest"],
+        properties={
+            "status": {"const": "PASS"},
+            "worktreeRoot": {"type": "string", "minLength": 1},
+            "allowedWriteRoots": {"type": "array", "items": {"type": "string", "minLength": 1}},
+            "preserveFailedAttempts": {"type": "boolean"},
+            "cleanupRequiresOperator": {"type": "boolean"},
+            "policyDigest": {"type": "string", "minLength": 64, "maxLength": 64},
+        },
+    ),
+    "agent-worktree-attempt-receipt.v1": _open_object_schema(
+        "agent-worktree-attempt-receipt.v1",
+        required=["schemaVersion", "status", "lineage", "taskId", "attempt", "policyDigest", "isolation", "outcome", "cleanup", "createdAt"],
+        properties={
+            "status": {"const": "PASS"},
+            "lineage": {"type": "object"},
+            "taskId": {"type": "string", "minLength": 1},
+            "attempt": {"type": "integer", "minimum": 1},
+            "policyDigest": {"type": "string", "minLength": 64, "maxLength": 64},
+            "isolation": {"type": "object"},
+            "outcome": {"enum": ["PASS", "FAILED", "BLOCKED"]},
+            "cleanup": {"type": "object"},
+            "createdAt": {"type": "string", "minLength": 1},
+            "receiptDigest": {"type": "string", "minLength": 64, "maxLength": 64},
+        },
+    ),
+    "agent-worktree-attempt-receipt-validation.v1": _open_object_schema(
+        "agent-worktree-attempt-receipt-validation.v1",
+        required=["schemaVersion", "status", "lineage", "taskId", "attempt", "outcome", "cleanupDecision", "receiptDigest"],
+        properties={
+            "status": {"const": "PASS"},
+            "lineage": {"type": "object"},
+            "taskId": {"type": "string", "minLength": 1},
+            "attempt": {"type": "integer", "minimum": 1},
+            "outcome": {"enum": ["PASS", "FAILED", "BLOCKED"]},
+            "cleanupDecision": {"enum": ["PRESERVE", "REMOVE"]},
+            "receiptDigest": {"type": "string", "minLength": 64, "maxLength": 64},
+        },
+    ),
+    "agent-runner-policy.v1": _open_object_schema(
+        "agent-runner-policy.v1",
+        required=[
+            "schemaVersion",
+            "maxAttemptsPerTask",
+            "maxReroutesPerTask",
+            "maxSplitsPerTask",
+            "maxBillableTokens",
+        ],
+        properties={
+            "maxAttemptsPerTask": {"type": "integer", "minimum": 0},
+            "maxReroutesPerTask": {"type": "integer", "minimum": 0},
+            "maxSplitsPerTask": {"type": "integer", "minimum": 0},
+            "maxBillableTokens": {"type": "integer", "minimum": 0},
+        },
+    ),
+    "agent-runner-state.v1": _open_object_schema(
+        "agent-runner-state.v1",
+        required=[
+            "schemaVersion",
+            "runnerRevision",
+            "status",
+            "currentTaskId",
+            "lineage",
+            "policy",
+            "counters",
+            "history",
+            "operations",
+        ],
+        properties={
+            "runnerRevision": {"type": "integer", "minimum": 1},
+            "status": {
+                "enum": [
+                    "READY",
+                    "ATTEMPTING",
+                    "VALIDATING",
+                    "REVIEWING",
+                    "WAITING_REMEDIATION",
+                    "WAITING_REROUTE",
+                    "WAITING_SPLIT",
+                    "BLOCKED",
+                    "STOPPED",
+                    "COMPLETE",
+                    "ABORTED",
+                ]
+            },
+            "currentTaskId": {"type": "string", "minLength": 1},
+            "lineage": {"type": "object"},
+            "policy": {"type": "object"},
+            "counters": {"type": "object"},
+            "history": {"type": "array", "items": {"type": "object"}},
+            "operations": {"type": "object"},
+            "stateDigest": {"type": "string", "minLength": 64, "maxLength": 64},
+        },
+    ),
+    "agent-runner-state-validation.v1": _open_object_schema(
+        "agent-runner-state-validation.v1",
+        required=["schemaVersion", "status", "runnerRevision", "runnerStatus", "lineage", "counters", "historyCount", "stateDigest"],
+        properties={
+            "status": {"const": "PASS"},
+            "runnerRevision": {"type": "integer", "minimum": 1},
+            "runnerStatus": {"type": "string", "minLength": 1},
+            "currentTaskId": {"type": ["string", "null"]},
+            "lineage": {"type": "object"},
+            "counters": {"type": "object"},
+            "historyCount": {"type": "integer", "minimum": 1},
+            "stateDigest": {"type": "string", "minLength": 64, "maxLength": 64},
+        },
+    ),
+    "agent-runner-transition-request.v1": _open_object_schema(
+        "agent-runner-transition-request.v1",
+        required=["schemaVersion", "operationId", "expectedRunnerRevision", "action", "reason"],
+        properties={
+            "operationId": {"type": "string", "minLength": 1},
+            "expectedRunnerRevision": {"type": "integer", "minimum": 1},
+            "action": {
+                "enum": [
+                    "attempt",
+                    "validate",
+                    "review",
+                    "accept",
+                    "remediate",
+                    "reroute",
+                    "split",
+                    "block",
+                    "abort",
+                ]
+            },
+            "taskId": {"type": "string", "minLength": 1},
+            "reason": {"type": "string", "minLength": 1},
+            "blockerCode": {"type": "string", "minLength": 1},
+            "evidenceIds": {"type": "array", "items": {"type": "string", "minLength": 1}},
+            "usage": {"type": "object"},
+            "patch": {"type": "object"},
+            "isolationReceipt": {"type": "object"},
+            "worktreePolicy": {"type": "object"},
+        },
+    ),
+    "agent-runner-transition-result.v1": _open_object_schema(
+        "agent-runner-transition-result.v1",
+        required=["schemaVersion", "status", "action", "runnerRevision", "runnerStatus", "transition", "allowedNextActions", "stateDigest", "resultDigest"],
+        properties={
+            "status": {"const": "PASS"},
+            "action": {"type": "string", "minLength": 1},
+            "runnerRevision": {"type": "integer", "minimum": 1},
+            "runnerStatus": {"type": "string", "minLength": 1},
+            "transition": {"type": "object"},
+            "allowedNextActions": {"type": "array", "items": {"type": "string", "minLength": 1}},
+            "stateDigest": {"type": "string", "minLength": 64, "maxLength": 64},
+            "resultDigest": {"type": "string", "minLength": 64, "maxLength": 64},
+        },
+    ),
+    "agent-runner-snapshot.v1": _open_object_schema(
+        "agent-runner-snapshot.v1",
+        required=["schemaVersion", "status", "runner", "budget", "lineage", "recentTransitions", "estimatedTokens", "snapshotDigest"],
+        properties={
+            "status": {"const": "PASS"},
+            "runner": {"type": "object"},
+            "budget": {"type": "object"},
+            "lineage": {"type": "object"},
+            "recentTransitions": {"type": "array", "items": {"type": "object"}},
+            "estimatedTokens": {"type": "integer", "minimum": 1},
+            "snapshotDigest": {"type": "string", "minLength": 64, "maxLength": 64},
+        },
+    ),
+    "agent-baseline-reconciliation-receipt.v1": _open_object_schema(
+        "agent-baseline-reconciliation-receipt.v1",
+        required=[
+            "schemaVersion",
+            "status",
+            "taskId",
+            "attempt",
+            "expectedBaseRevision",
+            "actualBaseRevision",
+            "evidenceIds",
+        ],
+        properties={
+            "status": {"const": "PASS"},
+            "taskId": {"type": "string", "minLength": 1},
+            "attempt": {"type": "integer", "minimum": 1},
+            "expectedBaseRevision": {"type": "string", "minLength": 1},
+            "actualBaseRevision": {"type": "string", "minLength": 1},
+            "evidenceIds": {"type": "array", "items": {"type": "string", "minLength": 1}},
+        },
+    ),
+    "agent-external-action-receipt.v1": _open_object_schema(
+        "agent-external-action-receipt.v1",
+        required=[
+            "schemaVersion",
+            "runId",
+            "packageId",
+            "planRevision",
+            "planDigest",
+            "sourceRevision",
+            "actionId",
+            "status",
+            "evidenceIds",
+        ],
+        properties={
+            "runId": {"type": "string", "minLength": 1},
+            "packageId": {"type": "string", "minLength": 1},
+            "planRevision": {"type": "integer", "minimum": 1},
+            "planDigest": {"type": "string", "minLength": 64, "maxLength": 64},
+            "sourceRevision": {"type": "string", "minLength": 1},
+            "actionId": {"type": "string", "minLength": 1},
+            "status": {"const": "PASS"},
+            "evidenceIds": {"type": "array", "items": {"type": "string", "minLength": 1}},
+        },
+    ),
+}
