+++ src/agent_lifecycle/contracts/schemas.py
@@ -5,1672 +5,40 @@ from __future__ import annotations
 from copy import deepcopy
 from typing import Any
 
+from agent_lifecycle.contracts.adapter_contract_schemas import ADAPTER_CONTRACT_SCHEMAS
+from agent_lifecycle.contracts.adapter_event_schemas import ADAPTER_EVENT_SCHEMAS
+from agent_lifecycle.contracts.context_model_schemas import CONTEXT_MODEL_SCHEMAS
+from agent_lifecycle.contracts.core_schemas import CORE_SCHEMAS
+from agent_lifecycle.contracts.evidence_import_schemas import EVIDENCE_IMPORT_SCHEMAS
+from agent_lifecycle.contracts.metric_schemas import METRIC_SCHEMAS
+from agent_lifecycle.contracts.plan_contract_schemas import PLAN_CONTRACT_SCHEMAS
+from agent_lifecycle.contracts.release_contract_schemas import RELEASE_CONTRACT_SCHEMAS
+from agent_lifecycle.contracts.review_quality_schemas import REVIEW_QUALITY_SCHEMAS
+from agent_lifecycle.contracts.runner_worktree_schemas import RUNNER_WORKTREE_SCHEMAS
+from agent_lifecycle.contracts.status_goal_schemas import STATUS_GOAL_SCHEMAS
+
 SCHEMA_INDEX_VERSION = "agent-lifecycle-schema-index.v1"
 
+_SCHEMA_GROUPS = (
+    CORE_SCHEMAS,
+    ADAPTER_EVENT_SCHEMAS,
+    REVIEW_QUALITY_SCHEMAS,
+    EVIDENCE_IMPORT_SCHEMAS,
+    STATUS_GOAL_SCHEMAS,
+    RUNNER_WORKTREE_SCHEMAS,
+    ADAPTER_CONTRACT_SCHEMAS,
+    CONTEXT_MODEL_SCHEMAS,
+    RELEASE_CONTRACT_SCHEMAS,
+    PLAN_CONTRACT_SCHEMAS,
+    METRIC_SCHEMAS,
+)
 
-def _open_object_schema(
-    schema_id: str,
-    *,
-    required: list[str],
-    properties: dict[str, Any] | None = None,
-) -> dict[str, Any]:
-    fields: dict[str, Any] = {"schemaVersion": {"const": schema_id}}
-    fields.update(properties or {})
-    return {
-        "$schema": "https://json-schema.org/draft/2020-12/schema",
-        "$id": schema_id,
-        "type": "object",
-        "additionalProperties": True,
-        "required": required,
-        "properties": fields,
-    }
-
-
-_SCHEMAS: dict[str, dict[str, Any]] = {
-    "agent-lifecycle-version.v1": {
-        "$schema": "https://json-schema.org/draft/2020-12/schema",
-        "$id": "agent-lifecycle-version.v1",
-        "type": "object",
-        "additionalProperties": False,
-        "required": ["schemaVersion", "version"],
-        "properties": {
-            "schemaVersion": {"const": "agent-lifecycle-version.v1"},
-            "version": {"type": "string", "minLength": 1},
-        },
-    },
-    "agent-lifecycle-schema-index.v1": {
-        "$schema": "https://json-schema.org/draft/2020-12/schema",
-        "$id": "agent-lifecycle-schema-index.v1",
-        "type": "object",
-        "additionalProperties": False,
-        "required": ["schemaVersion", "schemas"],
-        "properties": {
-            "schemaVersion": {"const": "agent-lifecycle-schema-index.v1"},
-            "schemas": {"type": "array", "items": {"type": "object"}},
-        },
-    },
-    "agent-lifecycle-error.v1": {
-        "$schema": "https://json-schema.org/draft/2020-12/schema",
-        "$id": "agent-lifecycle-error.v1",
-        "type": "object",
-        "additionalProperties": False,
-        "required": ["schemaVersion", "code", "message", "details"],
-        "properties": {
-            "schemaVersion": {"const": "agent-lifecycle-error.v1"},
-            "code": {"type": "string", "minLength": 1},
-            "message": {"type": "string", "minLength": 1},
-            "details": {"type": "object"},
-        },
-    },
-    "agent-host-operation-request.v1": {
-        "$schema": "https://json-schema.org/draft/2020-12/schema",
-        "$id": "agent-host-operation-request.v1",
-        "type": "object",
-        "additionalProperties": False,
-        "required": [
-            "schemaVersion",
-            "operationId",
-            "capability",
-            "inputs",
-            "outputs",
-            "constraints",
-        ],
-        "properties": {
-            "schemaVersion": {"const": "agent-host-operation-request.v1"},
-            "operationId": {"type": "string", "minLength": 1},
-            "capability": {"type": "string", "minLength": 1},
-            "inputs": {"type": "object"},
-            "outputs": {"type": "array", "items": {"type": "object"}},
-            "constraints": {"type": "object"},
-            "modelRoute": {"type": "object"},
-        },
-    },
-    "agent-host-operation-receipt.v1": {
-        "$schema": "https://json-schema.org/draft/2020-12/schema",
-        "$id": "agent-host-operation-receipt.v1",
-        "type": "object",
-        "additionalProperties": False,
-        "required": [
-            "schemaVersion",
-            "operationId",
-            "capability",
-            "status",
-            "outputs",
-            "usage",
-        ],
-        "properties": {
-            "schemaVersion": {"const": "agent-host-operation-receipt.v1"},
-            "operationId": {"type": "string", "minLength": 1},
-            "capability": {"type": "string", "minLength": 1},
-            "status": {"enum": ["PASS", "FAIL", "BLOCKED"]},
-            "outputs": {"type": "array", "items": {"type": "object"}},
-            "usage": {"type": "object"},
-        },
-    },
-    "agent-adapter-event.v1": {
-        "$schema": "https://json-schema.org/draft/2020-12/schema",
-        "$id": "agent-adapter-event.v1",
-        "type": "object",
-        "additionalProperties": False,
-        "required": [
-            "schemaVersion",
-            "eventId",
-            "host",
-            "adapterId",
-            "runId",
-            "taskId",
-            "operationId",
-            "sequence",
-            "eventType",
-            "status",
-            "recordedAt",
-            "payload",
-        ],
-        "properties": {
-            "schemaVersion": {"const": "agent-adapter-event.v1"},
-            "eventId": {"type": "string", "minLength": 1},
-            "host": {"type": "string", "minLength": 1},
-            "adapterId": {"type": "string", "minLength": 1},
-            "runId": {"type": "string", "minLength": 1},
-            "taskId": {"type": "string", "minLength": 1},
-            "operationId": {"type": "string", "minLength": 1},
-            "sequence": {"type": "integer", "minimum": 1},
-            "eventType": {
-                "enum": [
-                    "session.started",
-                    "task.launched",
-                    "command.completed",
-                    "writes.summarized",
-                    "usage.reported",
-                    "task.blocked",
-                    "task.completed",
-                ]
-            },
-            "status": {"enum": ["INFO", "PASS", "FAIL", "BLOCKED"]},
-            "recordedAt": {"type": "string", "minLength": 1},
-            "payload": {"type": "object"},
-        },
-    },
-    "agent-adapter-event-stream-validation.v1": _open_object_schema(
-        "agent-adapter-event-stream-validation.v1",
-        required=["schemaVersion", "status", "eventCount", "eventTypes", "blockers"],
-        properties={
-            "status": {"enum": ["PASS", "FAIL"]},
-            "host": {"type": ["string", "null"]},
-            "adapterId": {"type": ["string", "null"]},
-            "runId": {"type": ["string", "null"]},
-            "taskId": {"type": ["string", "null"]},
-            "eventCount": {"type": "integer", "minimum": 0},
-            "eventTypes": {"type": "array", "items": {"type": "string"}},
-            "terminalEvent": {"type": ["string", "null"]},
-            "blockers": {"type": "array", "items": {"type": "object"}},
-        },
-    ),
-    "agent-adapter-event-stream-receipt.v1": _open_object_schema(
-        "agent-adapter-event-stream-receipt.v1",
-        required=[
-            "schemaVersion",
-            "status",
-            "adapterId",
-            "host",
-            "runId",
-            "taskId",
-            "operationId",
-            "producer",
-            "descriptorDigest",
-            "eventStreamDigest",
-            "eventCount",
-            "eventTypes",
-            "terminalEvent",
-            "emittedAt",
-            "productionPromotionClaimed",
-        ],
-        properties={
-            "status": {"const": "PASS"},
-            "adapterId": {"type": "string", "minLength": 1},
-            "host": {"type": "string", "minLength": 1},
-            "runId": {"type": "string", "minLength": 1},
-            "taskId": {"type": "string", "minLength": 1},
-            "operationId": {"type": "string", "minLength": 1},
-            "producer": {"type": "object"},
-            "descriptorDigest": {"type": "string", "minLength": 64, "maxLength": 64},
-            "eventStreamDigest": {"type": "string", "minLength": 64, "maxLength": 64},
-            "eventCount": {"type": "integer", "minimum": 1},
-            "eventTypes": {"type": "array", "items": {"type": "string"}},
-            "terminalEvent": {"type": "string", "minLength": 1},
-            "emittedAt": {"type": "string", "minLength": 1},
-            "productionPromotionClaimed": {"const": False},
-        },
-    ),
-    "agent-adapter-event-capture-validation.v1": _open_object_schema(
-        "agent-adapter-event-capture-validation.v1",
-        required=["schemaVersion", "status", "adapterId", "host", "declaredEventCapture", "eventCount", "blockers", "productionPromotionClaimed"],
-        properties={
-            "status": {"enum": ["PASS", "FAIL"]},
-            "adapterId": {"type": ["string", "null"]},
-            "host": {"type": ["string", "null"]},
-            "declaredEventCapture": {"type": "boolean"},
-            "eventCount": {"type": "integer", "minimum": 0},
-            "blockers": {"type": "array", "items": {"type": "object"}},
-            "productionPromotionClaimed": {"const": False},
-        },
-    ),
-    "agent-review-verdict.v1": _open_object_schema(
-        "agent-review-verdict.v1",
-        required=["schemaVersion", "overall", "dimensions", "routing"],
-        properties={
-            "overall": {"enum": ["ACCEPTED", "REWORK", "CONTRACT_CHANGE", "BLOCKED"]},
-            "dimensions": {"type": "object"},
-            "routing": {"type": "object"},
-        },
-    ),
-    "agent-review-verdict-validation.v1": _open_object_schema(
-        "agent-review-verdict-validation.v1",
-        required=["schemaVersion", "status", "overall", "failingDimensions", "warningDimensions", "nextAction", "blockers"],
-        properties={
-            "status": {"enum": ["PASS", "FAIL"]},
-            "overall": {"type": ["string", "null"]},
-            "failingDimensions": {"type": "array", "items": {"type": "string"}},
-            "warningDimensions": {"type": "array", "items": {"type": "string"}},
-            "nextAction": {"type": ["string", "null"]},
-            "blockers": {"type": "array", "items": {"type": "object"}},
-        },
-    ),
-    "agent-review-routing-summary.v1": _open_object_schema(
-        "agent-review-routing-summary.v1",
-        required=["schemaVersion", "status", "overall", "nextAction", "failingDimensions", "warningDimensions", "dimensionStatus", "verdictDigest"],
-        properties={
-            "status": {"enum": ["PASS", "FAIL"]},
-            "overall": {"type": ["string", "null"]},
-            "nextAction": {"type": ["string", "null"]},
-            "failingDimensions": {"type": "array", "items": {"type": "string"}},
-            "warningDimensions": {"type": "array", "items": {"type": "string"}},
-            "dimensionStatus": {"type": "object"},
-            "verdictDigest": {"type": "string", "minLength": 64, "maxLength": 64},
-        },
-    ),
-    "agent-optional-quality-pack.v1": _open_object_schema(
-        "agent-optional-quality-pack.v1",
-        required=[
-            "schemaVersion",
-            "packId",
-            "status",
-            "enabledByDefault",
-            "activationMode",
-            "commands",
-            "productionPromotionClaimed",
-        ],
-        properties={
-            "packId": {"type": "string", "minLength": 1},
-            "status": {"const": "OPTIONAL"},
-            "enabledByDefault": {"const": False},
-            "activationMode": {"const": "opt-in"},
-            "commands": {"type": "array", "items": {"type": "object"}},
-            "productionPromotionClaimed": {"const": False},
-        },
-    ),
-    "agent-optional-quality-pack-validation.v1": _open_object_schema(
-        "agent-optional-quality-pack-validation.v1",
-        required=["schemaVersion", "status", "packId", "commandCount", "defaultEnabled", "blockers", "productionPromotionClaimed"],
-        properties={
-            "status": {"enum": ["PASS", "FAIL"]},
-            "packId": {"type": ["string", "null"]},
-            "commandCount": {"type": "integer", "minimum": 0},
-            "defaultEnabled": {"type": ["boolean", "null"]},
-            "blockers": {"type": "array", "items": {"type": "object"}},
-            "productionPromotionClaimed": {"const": False},
-        },
-    ),
-    "agent-behavior-check-fixture.v1": _open_object_schema(
-        "agent-behavior-check-fixture.v1",
-        required=["schemaVersion", "fixtureId", "expectedOutcome", "signals"],
-        properties={
-            "fixtureId": {"type": "string", "minLength": 1},
-            "expectedOutcome": {"enum": ["PASS", "FAIL", "BLOCKED"]},
-            "signals": {"type": "object"},
-        },
-    ),
-    "agent-behavior-check-run.v1": _open_object_schema(
-        "agent-behavior-check-run.v1",
-        required=[
-            "schemaVersion",
-            "status",
-            "packId",
-            "fixtureCount",
-            "passedExpectationCount",
-            "failedExpectationCount",
-            "checks",
-            "blockers",
-            "productionPromotionClaimed",
-        ],
-        properties={
-            "status": {"enum": ["PASS", "FAIL"]},
-            "packId": {"type": ["string", "null"]},
-            "fixtureCount": {"type": "integer", "minimum": 0},
-            "passedExpectationCount": {"type": "integer", "minimum": 0},
-            "failedExpectationCount": {"type": "integer", "minimum": 0},
-            "checks": {"type": "array", "items": {"type": "object"}},
-            "blockers": {"type": "array", "items": {"type": "object"}},
-            "productionPromotionClaimed": {"const": False},
-        },
-    ),
-    "agent-evidence-index.v1": _open_object_schema(
-        "agent-evidence-index.v1",
-        required=[
-            "schemaVersion",
-            "status",
-            "sourceOfTruth",
-            "rebuildable",
-            "enabledByDefault",
-            "activationMode",
-            "resourceCaps",
-            "artifactCount",
-            "entries",
-            "blockers",
-            "indexDigest",
-        ],
-        properties={
-            "status": {"enum": ["PASS", "FAIL"]},
-            "sourceOfTruth": {"const": False},
-            "rebuildable": {"const": True},
-            "enabledByDefault": {"const": False},
-            "activationMode": {"const": "explicit-command"},
-            "resourceCaps": {"type": "object"},
-            "artifactCount": {"type": "integer", "minimum": 0},
-            "entries": {"type": "array", "items": {"type": "object"}},
-            "blockers": {"type": "array", "items": {"type": "object"}},
-            "indexDigest": {"type": "string", "minLength": 64, "maxLength": 64},
-        },
-    ),
-    "agent-evidence-index-validation.v1": _open_object_schema(
-        "agent-evidence-index-validation.v1",
-        required=["schemaVersion", "status", "artifactCount", "blockers", "indexDigest", "validationDigest"],
-        properties={
-            "status": {"enum": ["PASS", "FAIL"]},
-            "artifactCount": {"type": "integer", "minimum": 0},
-            "blockers": {"type": "array", "items": {"type": "object"}},
-            "indexDigest": {"type": ["string", "null"], "minLength": 64, "maxLength": 64},
-            "validationDigest": {"type": "string", "minLength": 64, "maxLength": 64},
-        },
-    ),
-    "agent-evidence-search-summary.v1": _open_object_schema(
-        "agent-evidence-search-summary.v1",
-        required=["schemaVersion", "status", "sourceOfTruth", "query", "resultCount", "results", "indexDigest", "summaryDigest"],
-        properties={
-            "status": {"enum": ["PASS", "FAIL"]},
-            "sourceOfTruth": {"const": False},
-            "query": {"type": "string"},
-            "resultCount": {"type": "integer", "minimum": 0},
-            "results": {"type": "array", "items": {"type": "object"}},
-            "indexDigest": {"type": ["string", "null"], "minLength": 64, "maxLength": 64},
-            "summaryDigest": {"type": "string", "minLength": 64, "maxLength": 64},
-        },
-    ),
-    "agent-planning-import-result.v1": _open_object_schema(
-        "agent-planning-import-result.v1",
-        required=[
-            "schemaVersion",
-            "status",
-            "enabledByDefault",
-            "activationMode",
-            "sourceTrusted",
-            "candidateLifecycleStatus",
-            "reviewGates",
-            "resourceCaps",
-            "source",
-            "requiresReview",
-            "auditRequired",
-            "freezeBlocked",
-            "blockers",
-            "importDigest",
-        ],
-        properties={
-            "status": {"enum": ["PASS", "FAIL"]},
-            "enabledByDefault": {"const": False},
-            "activationMode": {"const": "explicit-command"},
-            "sourceTrusted": {"const": False},
-            "candidateLifecycleStatus": {"enum": ["DRAFT_REQUIRES_REVIEW", "BLOCKED"]},
-            "reviewGates": {"type": "array", "items": {"type": "string"}},
-            "resourceCaps": {"type": "object"},
-            "source": {"type": "object"},
-            "candidatePlan": {"type": ["object", "null"]},
-            "requiresReview": {"const": True},
-            "auditRequired": {"const": True},
-            "freezeBlocked": {"const": True},
-            "blockers": {"type": "array", "items": {"type": "object"}},
-            "importDigest": {"type": "string", "minLength": 64, "maxLength": 64},
-        },
-    ),
-    "agent-planning-import-validation.v1": _open_object_schema(
-        "agent-planning-import-validation.v1",
-        required=["schemaVersion", "status", "freezeBlocked", "auditRequired", "requiresReview", "blockers", "validationDigest"],
-        properties={
-            "status": {"enum": ["PASS", "FAIL"]},
-            "candidateDigest": {"type": ["string", "null"], "minLength": 64, "maxLength": 64},
-            "freezeBlocked": {"type": "boolean"},
-            "auditRequired": {"type": "boolean"},
-            "requiresReview": {"type": "boolean"},
-            "blockers": {"type": "array", "items": {"type": "object"}},
-            "validationDigest": {"type": "string", "minLength": 64, "maxLength": 64},
-        },
-    ),
-    "agent-skill-improvement-proposal.v1": _open_object_schema(
-        "agent-skill-improvement-proposal.v1",
-        required=[
-            "schemaVersion",
-            "proposalId",
-            "affectedSkill",
-            "status",
-            "rationale",
-            "expectedBehavior",
-            "requiredTests",
-            "requiresReview",
-            "autoApply",
-            "applied",
-        ],
-        properties={
-            "proposalId": {"type": "string", "minLength": 1},
-            "affectedSkill": {"type": "string", "minLength": 1},
-            "status": {"const": "PROPOSED"},
-            "rationale": {"type": "string", "minLength": 1},
-            "expectedBehavior": {"type": "string", "minLength": 1},
-            "requiredTests": {"type": "array", "items": {"type": "string", "minLength": 1}},
-            "requiresReview": {"const": True},
-            "autoApply": {"const": False},
-            "applied": {"const": False},
-        },
-    ),
-    "agent-skill-improvement-proposal-validation.v1": _open_object_schema(
-        "agent-skill-improvement-proposal-validation.v1",
-        required=["schemaVersion", "status", "proposalId", "requiresReview", "autoApply", "applied", "blockers", "proposalDigest", "validationDigest"],
-        properties={
-            "status": {"enum": ["PASS", "FAIL"]},
-            "proposalId": {"type": ["string", "null"]},
-            "requiresReview": {"type": "boolean"},
-            "autoApply": {"type": ["boolean", "null"]},
-            "applied": {"type": ["boolean", "null"]},
-            "blockers": {"type": "array", "items": {"type": "object"}},
-            "proposalDigest": {"type": "string", "minLength": 64, "maxLength": 64},
-            "validationDigest": {"type": "string", "minLength": 64, "maxLength": 64},
-        },
-    ),
-    "agent-diagnostic-bundle.v1": _open_object_schema(
-        "agent-diagnostic-bundle.v1",
-        required=[
-            "schemaVersion",
-            "status",
-            "projectRoot",
-            "sourceOfTruth",
-            "redacted",
-            "artifacts",
-            "blockers",
-            "productionPromotionClaimed",
-        ],
-        properties={
-            "status": {"enum": ["PASS", "FAIL"]},
-            "projectRoot": {"const": "<checkout>"},
-            "sourceOfTruth": {"const": False},
-            "redacted": {"const": True},
-            "artifacts": {"type": "array", "items": {"type": "object"}},
-            "blockers": {"type": "array", "items": {"type": "object"}},
-            "productionPromotionClaimed": {"const": False},
-        },
-    ),
-    "agent-readonly-status-view.v1": _open_object_schema(
-        "agent-readonly-status-view.v1",
-        required=[
-            "schemaVersion",
-            "status",
-            "sourceOfTruth",
-            "projectRoot",
-            "itemCount",
-            "items",
-            "blockers",
-            "productionPromotionClaimed",
-        ],
-        properties={
-            "status": {"enum": ["PASS", "FAIL"]},
-            "sourceOfTruth": {"const": False},
-            "projectRoot": {"const": "<checkout>"},
-            "itemCount": {"type": "integer", "minimum": 0},
-            "items": {"type": "array", "items": {"type": "object"}},
-            "blockers": {"type": "array", "items": {"type": "object"}},
-            "productionPromotionClaimed": {"const": False},
-        },
-    ),
-    "agent-completion-signal.v1": {
-        "$schema": "https://json-schema.org/draft/2020-12/schema",
-        "$id": "agent-completion-signal.v1",
-        "type": "object",
-        "additionalProperties": True,
-        "required": [
-            "schemaVersion",
-            "runId",
-            "packageId",
-            "planRevision",
-            "planDigest",
-            "sourceRevision",
-            "status",
-            "evidenceIds",
-            "verifier",
-        ],
-        "properties": {
-            "schemaVersion": {"const": "agent-completion-signal.v1"},
-            "runId": {"type": "string", "minLength": 1},
-            "packageId": {"type": "string", "minLength": 1},
-            "planRevision": {"type": "integer", "minimum": 1},
-            "planDigest": {"type": "string", "minLength": 64, "maxLength": 64},
-            "sourceRevision": {"type": "string", "minLength": 1},
-            "status": {"enum": ["PASS", "WAIVED", "FAIL"]},
-            "evidenceIds": {"type": "array", "items": {"type": "string", "minLength": 1}},
-            "verifier": {"type": "object"},
-            "waiver": {"type": "object"},
-        },
-    },
-    "agent-completion-signal-validation.v1": _open_object_schema(
-        "agent-completion-signal-validation.v1",
-        required=["schemaVersion", "status", "signalStatus", "signalDigest", "waived"],
-        properties={
-            "status": {"const": "PASS"},
-            "signalStatus": {"enum": ["PASS", "WAIVED"]},
-            "signalDigest": {"type": "string", "minLength": 64, "maxLength": 64},
-            "waived": {"type": "boolean"},
-        },
-    ),
-    "agent-completion-check.v1": _open_object_schema(
-        "agent-completion-check.v1",
-        required=["schemaVersion", "checkId", "kind", "description", "receiptPath", "requiredEvidenceIds"],
-        properties={
-            "checkId": {"type": "string", "minLength": 1},
-            "kind": {"enum": ["verification", "external-action"]},
-            "description": {"type": "string", "minLength": 1},
-            "receiptPath": {"type": "string", "minLength": 1},
-            "requiredEvidenceIds": {"type": "array", "items": {"type": "string", "minLength": 1}},
-        },
-    ),
-    "agent-completion-check-validation.v1": _open_object_schema(
-        "agent-completion-check-validation.v1",
-        required=["schemaVersion", "checkId", "kind", "receiptPath", "requiredEvidenceIds", "checkDigest"],
-        properties={
-            "checkId": {"type": "string", "minLength": 1},
-            "kind": {"enum": ["verification", "external-action"]},
-            "receiptPath": {"type": "string", "minLength": 1},
-            "requiredEvidenceIds": {"type": "array", "items": {"type": "string", "minLength": 1}},
-            "checkDigest": {"type": "string", "minLength": 64, "maxLength": 64},
-        },
-    ),
-    "agent-completion-check-receipt.v1": _open_object_schema(
-        "agent-completion-check-receipt.v1",
-        required=[
-            "schemaVersion",
-            "checkId",
-            "runId",
-            "packageId",
-            "planRevision",
-            "planDigest",
-            "sourceRevision",
-            "status",
-            "evidenceIds",
-            "verifier",
-        ],
-        properties={
-            "checkId": {"type": "string", "minLength": 1},
-            "runId": {"type": "string", "minLength": 1},
-            "packageId": {"type": "string", "minLength": 1},
-            "planRevision": {"type": "integer", "minimum": 1},
-            "planDigest": {"type": "string", "minLength": 64, "maxLength": 64},
-            "sourceRevision": {"type": "string", "minLength": 1},
-            "status": {"enum": ["PASS", "FAIL"]},
-            "evidenceIds": {"type": "array", "items": {"type": "string", "minLength": 1}},
-            "verifier": {"type": "object"},
-            "externalActionReceipt": {"type": "object"},
-        },
-    ),
-    "agent-completion-check-receipt-validation.v1": _open_object_schema(
-        "agent-completion-check-receipt-validation.v1",
-        required=["schemaVersion", "status", "checkId", "checkKind", "evidenceIds", "receiptDigest", "checkDigest"],
-        properties={
-            "status": {"const": "PASS"},
-            "checkId": {"type": "string", "minLength": 1},
-            "checkKind": {"enum": ["verification", "external-action"]},
-            "evidenceIds": {"type": "array", "items": {"type": "string", "minLength": 1}},
-            "receiptDigest": {"type": "string", "minLength": 64, "maxLength": 64},
-            "checkDigest": {"type": "string", "minLength": 64, "maxLength": 64},
-        },
-    ),
-    "agent-goal-record.v1": _open_object_schema(
-        "agent-goal-record.v1",
-        required=[
-            "schemaVersion",
-            "goalId",
-            "userIntent",
-            "ownerOutcome",
-            "constraints",
-            "status",
-            "lineage",
-            "evidenceIds",
-            "updatedAt",
-        ],
-        properties={
-            "goalId": {"type": "string", "minLength": 1},
-            "userIntent": {"type": "string", "minLength": 1},
-            "ownerOutcome": {"type": "string", "minLength": 1},
-            "constraints": {"type": "array", "items": {"type": "string", "minLength": 1}},
-            "status": {"enum": ["ACTIVE", "BLOCKED", "READY_FOR_FINALIZATION", "COMPLETE"]},
-            "lineage": {"type": "object"},
-            "completionCheck": {"type": "object"},
-            "evidenceIds": {"type": "array", "items": {"type": "string", "minLength": 1}},
-            "updatedAt": {"type": "string", "minLength": 1},
-        },
-    ),
-    "agent-goal-record-validation.v1": _open_object_schema(
-        "agent-goal-record-validation.v1",
-        required=["schemaVersion", "status", "goalId", "goalStatus", "lineage", "stateRevision", "goalDigest"],
-        properties={
-            "status": {"const": "PASS"},
-            "goalId": {"type": "string", "minLength": 1},
-            "goalStatus": {"enum": ["ACTIVE", "BLOCKED", "READY_FOR_FINALIZATION", "COMPLETE"]},
-            "lineage": {"type": "object"},
-            "stateRevision": {"type": "integer", "minimum": 1},
-            "goalDigest": {"type": "string", "minLength": 64, "maxLength": 64},
-        },
-    ),
-    "agent-objective-snapshot.v1": _open_object_schema(
-        "agent-objective-snapshot.v1",
-        required=["schemaVersion", "status", "goal", "lifecycle", "estimatedTokens", "snapshotDigest"],
-        properties={
-            "status": {"const": "PASS"},
-            "goal": {"type": "object"},
-            "lifecycle": {"type": "object"},
-            "completionCheck": {"type": ["object", "null"]},
-            "estimatedTokens": {"type": "integer", "minimum": 1},
-            "snapshotDigest": {"type": "string", "minLength": 64, "maxLength": 64},
-        },
-    ),
-    "agent-follow-up-register.v1": _open_object_schema(
-        "agent-follow-up-register.v1",
-        required=["schemaVersion", "lineage", "items"],
-        properties={
-            "lineage": {"type": "object"},
-            "items": {"type": "array", "items": {"type": "object"}},
-            "updatedAt": {"type": "string", "minLength": 1},
-        },
-    ),
-    "agent-follow-up-register-validation.v1": _open_object_schema(
-        "agent-follow-up-register-validation.v1",
-        required=[
-            "schemaVersion",
-            "status",
-            "registerDigest",
-            "lineage",
-            "itemCount",
-            "openItemIds",
-            "finalizationBlockers",
-            "items",
-            "validationDigest",
-        ],
-        properties={
-            "status": {"const": "PASS"},
-            "registerDigest": {"type": "string", "minLength": 64, "maxLength": 64},
-            "lineage": {"type": "object"},
-            "itemCount": {"type": "integer", "minimum": 0},
-            "openItemIds": {"type": "array", "items": {"type": "string", "minLength": 1}},
-            "finalizationBlockers": {"type": "array", "items": {"type": "object"}},
-            "items": {"type": "array", "items": {"type": "object"}},
-            "validationDigest": {"type": "string", "minLength": 64, "maxLength": 64},
-        },
-    ),
-    "agent-follow-up-close-result.v1": _open_object_schema(
-        "agent-follow-up-close-result.v1",
-        required=["schemaVersion", "status", "itemId", "registerDigest"],
-        properties={
-            "status": {"const": "PASS"},
-            "itemId": {"type": "string", "minLength": 1},
-            "registerDigest": {"type": "string", "minLength": 64, "maxLength": 64},
-        },
-    ),
-    "agent-follow-up-summary.v1": _open_object_schema(
-        "agent-follow-up-summary.v1",
-        required=["schemaVersion", "status", "lineage", "registerDigest", "counts", "openItems", "finalizationBlockers", "estimatedTokens", "summaryDigest"],
-        properties={
-            "status": {"const": "PASS"},
-            "lineage": {"type": "object"},
-            "registerDigest": {"type": "string", "minLength": 64, "maxLength": 64},
-            "counts": {"type": "object"},
-            "openItems": {"type": "array", "items": {"type": "object"}},
-            "finalizationBlockers": {"type": "array", "items": {"type": "object"}},
-            "estimatedTokens": {"type": "integer", "minimum": 1},
-            "summaryDigest": {"type": "string", "minLength": 64, "maxLength": 64},
-        },
-    ),
-    "agent-worktree-isolation-policy.v1": _open_object_schema(
-        "agent-worktree-isolation-policy.v1",
-        required=["schemaVersion", "worktreeRoot", "preserveFailedAttempts", "cleanupRequiresOperator"],
-        properties={
-            "worktreeRoot": {"type": "string", "minLength": 1},
-            "allowedWriteRoots": {"type": "array", "items": {"type": "string", "minLength": 1}},
-            "preserveFailedAttempts": {"type": "boolean"},
-            "cleanupRequiresOperator": {"type": "boolean"},
-        },
-    ),
-    "agent-worktree-isolation-policy-validation.v1": _open_object_schema(
-        "agent-worktree-isolation-policy-validation.v1",
-        required=["schemaVersion", "status", "worktreeRoot", "allowedWriteRoots", "preserveFailedAttempts", "cleanupRequiresOperator", "policyDigest"],
-        properties={
-            "status": {"const": "PASS"},
-            "worktreeRoot": {"type": "string", "minLength": 1},
-            "allowedWriteRoots": {"type": "array", "items": {"type": "string", "minLength": 1}},
-            "preserveFailedAttempts": {"type": "boolean"},
-            "cleanupRequiresOperator": {"type": "boolean"},
-            "policyDigest": {"type": "string", "minLength": 64, "maxLength": 64},
-        },
-    ),
-    "agent-worktree-attempt-receipt.v1": _open_object_schema(
-        "agent-worktree-attempt-receipt.v1",
-        required=["schemaVersion", "status", "lineage", "taskId", "attempt", "policyDigest", "isolation", "outcome", "cleanup", "createdAt"],
-        properties={
-            "status": {"const": "PASS"},
-            "lineage": {"type": "object"},
-            "taskId": {"type": "string", "minLength": 1},
-            "attempt": {"type": "integer", "minimum": 1},
-            "policyDigest": {"type": "string", "minLength": 64, "maxLength": 64},
-            "isolation": {"type": "object"},
-            "outcome": {"enum": ["PASS", "FAILED", "BLOCKED"]},
-            "cleanup": {"type": "object"},
-            "createdAt": {"type": "string", "minLength": 1},
-            "receiptDigest": {"type": "string", "minLength": 64, "maxLength": 64},
-        },
-    ),
-    "agent-worktree-attempt-receipt-validation.v1": _open_object_schema(
-        "agent-worktree-attempt-receipt-validation.v1",
-        required=["schemaVersion", "status", "lineage", "taskId", "attempt", "outcome", "cleanupDecision", "receiptDigest"],
-        properties={
-            "status": {"const": "PASS"},
-            "lineage": {"type": "object"},
-            "taskId": {"type": "string", "minLength": 1},
-            "attempt": {"type": "integer", "minimum": 1},
-            "outcome": {"enum": ["PASS", "FAILED", "BLOCKED"]},
-            "cleanupDecision": {"enum": ["PRESERVE", "REMOVE"]},
-            "receiptDigest": {"type": "string", "minLength": 64, "maxLength": 64},
-        },
-    ),
-    "agent-runner-policy.v1": _open_object_schema(
-        "agent-runner-policy.v1",
-        required=[
-            "schemaVersion",
-            "maxAttemptsPerTask",
-            "maxReroutesPerTask",
-            "maxSplitsPerTask",
-            "maxBillableTokens",
-        ],
-        properties={
-            "maxAttemptsPerTask": {"type": "integer", "minimum": 0},
-            "maxReroutesPerTask": {"type": "integer", "minimum": 0},
-            "maxSplitsPerTask": {"type": "integer", "minimum": 0},
-            "maxBillableTokens": {"type": "integer", "minimum": 0},
-        },
-    ),
-    "agent-runner-state.v1": _open_object_schema(
-        "agent-runner-state.v1",
-        required=[
-            "schemaVersion",
-            "runnerRevision",
-            "status",
-            "currentTaskId",
-            "lineage",
-            "policy",
-            "counters",
-            "history",
-            "operations",
-        ],
-        properties={
-            "runnerRevision": {"type": "integer", "minimum": 1},
-            "status": {
-                "enum": [
-                    "READY",
-                    "ATTEMPTING",
-                    "VALIDATING",
-                    "REVIEWING",
-                    "WAITING_REMEDIATION",
-                    "WAITING_REROUTE",
-                    "WAITING_SPLIT",
-                    "BLOCKED",
-                    "STOPPED",
-                    "COMPLETE",
-                    "ABORTED",
-                ]
-            },
-            "currentTaskId": {"type": "string", "minLength": 1},
-            "lineage": {"type": "object"},
-            "policy": {"type": "object"},
-            "counters": {"type": "object"},
-            "history": {"type": "array", "items": {"type": "object"}},
-            "operations": {"type": "object"},
-            "stateDigest": {"type": "string", "minLength": 64, "maxLength": 64},
-        },
-    ),
-    "agent-runner-state-validation.v1": _open_object_schema(
-        "agent-runner-state-validation.v1",
-        required=["schemaVersion", "status", "runnerRevision", "runnerStatus", "lineage", "counters", "historyCount", "stateDigest"],
-        properties={
-            "status": {"const": "PASS"},
-            "runnerRevision": {"type": "integer", "minimum": 1},
-            "runnerStatus": {"type": "string", "minLength": 1},
-            "currentTaskId": {"type": ["string", "null"]},
-            "lineage": {"type": "object"},
-            "counters": {"type": "object"},
-            "historyCount": {"type": "integer", "minimum": 1},
-            "stateDigest": {"type": "string", "minLength": 64, "maxLength": 64},
-        },
-    ),
-    "agent-runner-transition-request.v1": _open_object_schema(
-        "agent-runner-transition-request.v1",
-        required=["schemaVersion", "operationId", "expectedRunnerRevision", "action", "reason"],
-        properties={
-            "operationId": {"type": "string", "minLength": 1},
-            "expectedRunnerRevision": {"type": "integer", "minimum": 1},
-            "action": {
-                "enum": [
-                    "attempt",
-                    "validate",
-                    "review",
-                    "accept",
-                    "remediate",
-                    "reroute",
-                    "split",
-                    "block",
-                    "abort",
-                ]
-            },
-            "taskId": {"type": "string", "minLength": 1},
-            "reason": {"type": "string", "minLength": 1},
-            "blockerCode": {"type": "string", "minLength": 1},
-            "evidenceIds": {"type": "array", "items": {"type": "string", "minLength": 1}},
-            "usage": {"type": "object"},
-            "patch": {"type": "object"},
-            "isolationReceipt": {"type": "object"},
-            "worktreePolicy": {"type": "object"},
-        },
-    ),
-    "agent-runner-transition-result.v1": _open_object_schema(
-        "agent-runner-transition-result.v1",
-        required=["schemaVersion", "status", "action", "runnerRevision", "runnerStatus", "transition", "allowedNextActions", "stateDigest", "resultDigest"],
-        properties={
-            "status": {"const": "PASS"},
-            "action": {"type": "string", "minLength": 1},
-            "runnerRevision": {"type": "integer", "minimum": 1},
-            "runnerStatus": {"type": "string", "minLength": 1},
-            "transition": {"type": "object"},
-            "allowedNextActions": {"type": "array", "items": {"type": "string", "minLength": 1}},
-            "stateDigest": {"type": "string", "minLength": 64, "maxLength": 64},
-            "resultDigest": {"type": "string", "minLength": 64, "maxLength": 64},
-        },
-    ),
-    "agent-runner-snapshot.v1": _open_object_schema(
-        "agent-runner-snapshot.v1",
-        required=["schemaVersion", "status", "runner", "budget", "lineage", "recentTransitions", "estimatedTokens", "snapshotDigest"],
-        properties={
-            "status": {"const": "PASS"},
-            "runner": {"type": "object"},
-            "budget": {"type": "object"},
-            "lineage": {"type": "object"},
-            "recentTransitions": {"type": "array", "items": {"type": "object"}},
-            "estimatedTokens": {"type": "integer", "minimum": 1},
-            "snapshotDigest": {"type": "string", "minLength": 64, "maxLength": 64},
-        },
-    ),
-    "agent-baseline-reconciliation-receipt.v1": _open_object_schema(
-        "agent-baseline-reconciliation-receipt.v1",
-        required=[
-            "schemaVersion",
-            "status",
-            "taskId",
-            "attempt",
-            "expectedBaseRevision",
-            "actualBaseRevision",
-            "evidenceIds",
-        ],
-        properties={
-            "status": {"const": "PASS"},
-            "taskId": {"type": "string", "minLength": 1},
-            "attempt": {"type": "integer", "minimum": 1},
-            "expectedBaseRevision": {"type": "string", "minLength": 1},
-            "actualBaseRevision": {"type": "string", "minLength": 1},
-            "evidenceIds": {"type": "array", "items": {"type": "string", "minLength": 1}},
-        },
-    ),
-    "agent-external-action-receipt.v1": _open_object_schema(
-        "agent-external-action-receipt.v1",
-        required=[
-            "schemaVersion",
-            "runId",
-            "packageId",
-            "planRevision",
-            "planDigest",
-            "sourceRevision",
-            "actionId",
-            "status",
-            "evidenceIds",
-        ],
-        properties={
-            "runId": {"type": "string", "minLength": 1},
-            "packageId": {"type": "string", "minLength": 1},
-            "planRevision": {"type": "integer", "minimum": 1},
-            "planDigest": {"type": "string", "minLength": 64, "maxLength": 64},
-            "sourceRevision": {"type": "string", "minLength": 1},
-            "actionId": {"type": "string", "minLength": 1},
-            "status": {"const": "PASS"},
-            "evidenceIds": {"type": "array", "items": {"type": "string", "minLength": 1}},
-        },
-    ),
-    "agent-host-adapter-validation.v1": _open_object_schema(
-        "agent-host-adapter-validation.v1",
-        required=["schemaVersion", "status", "adapterId", "host", "maturity", "blockers"],
-        properties={
-            "status": {"enum": ["PASS", "FAIL"]},
-            "adapterId": {"type": ["string", "null"]},
-            "host": {"type": ["string", "null"]},
-            "maturity": {"type": ["string", "null"]},
-            "blockers": {"type": "array", "items": {"type": "object"}},
-        },
-    ),
-    "agent-readiness-report.v1": _open_object_schema(
-        "agent-readiness-report.v1",
-        required=["schemaVersion", "status", "summary", "checks", "adapters", "evidence"],
-        properties={
-            "status": {"enum": ["PASS", "WARN", "FAIL"]},
-            "summary": {"type": "object"},
-            "checks": {"type": "array", "items": {"type": "object"}},
-            "adapters": {"type": "array", "items": {"type": "object"}},
-            "evidence": {"type": "object"},
-            "productionPromotionClaimed": {"const": False},
-            "maturityChangesClaimed": {"const": False},
-        },
-    ),
-    "agent-adapter-install-plan.v1": _open_object_schema(
-        "agent-adapter-install-plan.v1",
-        required=["schemaVersion", "status", "host", "maturity", "files", "commands"],
-        properties={
-            "status": {"const": "DRY_RUN"},
-            "host": {"type": "string", "minLength": 1},
-            "maturity": {"enum": ["EXPERIMENTAL", "VERIFIED"]},
-            "files": {"type": "array", "items": {"type": "object"}},
-            "commands": {"type": "array", "items": {"type": "object"}},
-            "writesStarted": {"const": False},
-            "productionPromotionClaimed": {"const": False},
-            "maturityChangeClaimed": {"const": False},
-        },
-    ),
-    "agent-host-adapter-inspection.v1": _open_object_schema(
-        "agent-host-adapter-inspection.v1",
-        required=[
-            "schemaVersion",
-            "status",
-            "adapterId",
-            "host",
-            "maturity",
-            "descriptorDigest",
-            "liveCallsStarted",
-            "productionPromotionClaimed",
-            "capabilities",
-            "checks",
-            "blockers",
-        ],
-        properties={
-            "status": {"enum": ["PASS", "FAIL"]},
-            "adapterId": {"type": ["string", "null"]},
-            "host": {"type": ["string", "null"]},
-            "maturity": {"type": ["string", "null"]},
-            "descriptorDigest": {"type": "string", "minLength": 64, "maxLength": 64},
-            "liveCallsStarted": {"const": False},
-            "productionPromotionClaimed": {"const": False},
-            "capabilities": {"type": "object"},
-            "checks": {"type": "array", "items": {"type": "object"}},
-            "blockers": {"type": "array", "items": {"type": "object"}},
-        },
-    ),
-    "agent-adapter-capability-manifest.v1": _open_object_schema(
-        "agent-adapter-capability-manifest.v1",
-        required=[
-            "schemaVersion",
-            "adapterId",
-            "host",
-            "maturity",
-            "descriptorDigest",
-            "unsupportedOperationPolicy",
-            "coreSemantics",
-            "capabilities",
-            "modelRouting",
-            "runtimeBoundary",
-            "promotion",
-        ],
-        properties={
-            "adapterId": {"type": ["string", "null"]},
-            "host": {"type": ["string", "null"]},
-            "maturity": {"type": ["string", "null"]},
-            "descriptorDigest": {"type": "string", "minLength": 64, "maxLength": 64},
-            "unsupportedOperationPolicy": {"const": "fail-closed"},
-            "coreSemantics": {"const": "delegated-to-agent-lifecycle-core"},
-            "capabilities": {"type": "array", "items": {"type": "object"}},
-            "modelRouting": {"type": "object"},
-            "runtimeBoundary": {"type": "object"},
-            "promotion": {"type": "object"},
-        },
-    ),
-    "agent-adapter-capability-manifest-validation.v1": _open_object_schema(
-        "agent-adapter-capability-manifest-validation.v1",
-        required=["schemaVersion", "status", "adapterId", "host", "maturity", "capabilityCount", "blockers"],
-        properties={
-            "status": {"enum": ["PASS", "FAIL"]},
-            "adapterId": {"type": ["string", "null"]},
-            "host": {"type": ["string", "null"]},
-            "maturity": {"type": ["string", "null"]},
-            "capabilityCount": {"type": "integer", "minimum": 0},
-            "blockers": {"type": "array", "items": {"type": "object"}},
-        },
-    ),
-    "agent-adapter-conformance-verification.v1": _open_object_schema(
-        "agent-adapter-conformance-verification.v1",
-        required=["schemaVersion", "status", "baseline", "hosts", "checks", "blockers", "productionPromotionClaimed"],
-        properties={
-            "status": {"enum": ["PASS", "FAIL"]},
-            "baseline": {"type": "object"},
-            "hosts": {"type": "array", "items": {"type": "string"}},
-            "checks": {"type": "array", "items": {"type": "object"}},
-            "blockers": {"type": "array", "items": {"type": "object"}},
-            "productionPromotionClaimed": {"const": False},
-        },
-    ),
-    "agent-small-context-profile.v1": {
-        "$schema": "https://json-schema.org/draft/2020-12/schema",
-        "$id": "agent-small-context-profile.v1",
-        "type": "object",
-        "additionalProperties": True,
-        "required": ["schemaVersion", "profileId", "defaultWindow", "windows", "requiredSummaryFields"],
-        "properties": {
-            "schemaVersion": {"const": "agent-small-context-profile.v1"},
-            "profileId": {"type": "string", "minLength": 1},
-            "defaultWindow": {"enum": ["4k-strict", "8k", "16k", "32k", "64k"]},
-            "windows": {"type": "object"},
-            "requiredSummaryFields": {"type": "array", "items": {"type": "string"}},
-        },
-    },
-    "agent-context-render-receipt.v1": {
-        "$schema": "https://json-schema.org/draft/2020-12/schema",
-        "$id": "agent-context-render-receipt.v1",
-        "type": "object",
-        "additionalProperties": True,
-        "required": ["schemaVersion", "status", "profileId", "window", "estimatedTokens", "checks", "envelopeDigest"],
-        "properties": {
-            "schemaVersion": {"const": "agent-context-render-receipt.v1"},
-            "status": {"enum": ["PASS", "FAIL"]},
-            "profileId": {"type": "string", "minLength": 1},
-            "window": {"enum": ["4k-strict", "8k", "16k", "32k", "64k"]},
-            "estimatedTokens": {"type": "object"},
-            "checks": {"type": "array", "items": {"type": "object"}},
-            "envelopeDigest": {"type": "string", "minLength": 64, "maxLength": 64},
-        },
-    },
-    "agent-sdd-tier-resolution.v1": {
-        "$schema": "https://json-schema.org/draft/2020-12/schema",
-        "$id": "agent-sdd-tier-resolution.v1",
-        "type": "object",
-        "additionalProperties": False,
-        "required": ["schemaVersion", "tier", "reasons", "requestDigest", "rules"],
-        "properties": {
-            "schemaVersion": {"const": "agent-sdd-tier-resolution.v1"},
-            "tier": {"enum": ["S0", "S1", "S2"]},
-            "reasons": {"type": "array", "items": {"type": "string"}},
-            "requestDigest": {"type": "string", "minLength": 64, "maxLength": 64},
-            "rules": {"type": "object"},
-        },
-    },
-    "agent-lifecycle-model-routing-profile.v1": {
-        "$schema": "https://json-schema.org/draft/2020-12/schema",
-        "$id": "agent-lifecycle-model-routing-profile.v1",
-        "type": "object",
-        "additionalProperties": True,
-        "required": ["schemaVersion", "profileId", "classes", "phaseRules"],
-        "properties": {
-            "schemaVersion": {"const": "agent-lifecycle-model-routing-profile.v1"},
-            "profileId": {"type": "string", "minLength": 1},
-            "classes": {"type": "array", "items": {"type": "string", "minLength": 1}},
-            "phaseRules": {"type": "object"},
-        },
-    },
-    "agent-lifecycle-host-model-profile.v1": {
-        "$schema": "https://json-schema.org/draft/2020-12/schema",
-        "$id": "agent-lifecycle-host-model-profile.v1",
-        "type": "object",
-        "additionalProperties": True,
-        "required": ["schemaVersion", "host", "profileId", "bindings"],
-        "properties": {
-            "schemaVersion": {"const": "agent-lifecycle-host-model-profile.v1"},
-            "host": {"type": "string", "minLength": 1},
-            "profileId": {"type": "string", "minLength": 1},
-            "bindings": {"type": "object"},
-        },
-    },
-    "agent-host-model-selection-profile.v1": {
-        "$schema": "https://json-schema.org/draft/2020-12/schema",
-        "$id": "agent-host-model-selection-profile.v1",
-        "type": "object",
-        "additionalProperties": True,
-        "required": [
-            "schemaVersion",
-            "host",
-            "profileId",
-            "budgetModeDefault",
-            "bindings",
-            "fallbackPolicies",
-            "redactionPolicy",
-        ],
-        "properties": {
-            "schemaVersion": {"const": "agent-host-model-selection-profile.v1"},
-            "host": {"type": "string", "minLength": 1},
-            "profileId": {"type": "string", "minLength": 1},
-            "budgetModeDefault": {"enum": ["metered", "subscription", "local"]},
-            "bindings": {"type": "object"},
-            "fallbackPolicies": {"type": "object"},
-            "redactionPolicy": {"type": "object"},
-        },
-    },
-    "agent-host-model-selection-receipt.v1": _open_object_schema(
-        "agent-host-model-selection-receipt.v1",
-        required=[
-            "schemaVersion",
-            "host",
-            "routeDecisionDigest",
-            "modelClass",
-            "profileDigest",
-            "bindingId",
-            "bindingDigest",
-            "providerModelHash",
-            "fallbackUsed",
-        ],
-        properties={
-            "host": {"type": "string", "minLength": 1},
-            "routeDecisionDigest": {"type": "string", "minLength": 64, "maxLength": 64},
-            "modelClass": {"type": "string", "minLength": 1},
-            "profileDigest": {"type": "string", "minLength": 64, "maxLength": 64},
-            "bindingId": {"type": "string", "minLength": 1},
-            "bindingDigest": {"type": "string", "minLength": 64, "maxLength": 64},
-            "providerModelHash": {"type": "string", "minLength": 64, "maxLength": 64},
-            "fallbackUsed": {"type": "boolean"},
-            "fallbackReason": {"type": ["string", "null"]},
-        },
-    ),
-    "agent-lifecycle-budget-exceeded-policy.v1": _open_object_schema(
-        "agent-lifecycle-budget-exceeded-policy.v1",
-        required=[
-            "schemaVersion",
-            "mode",
-            "allowedActions",
-            "forbidDowngradeForCriticalReview",
-            "maxAutoReroutesPerTask",
-            "budgetModes",
-        ],
-        properties={
-            "mode": {"enum": ["manual", "auto"]},
-            "allowedActions": {
-                "type": "array",
-                "items": {
-                    "enum": [
-                        "continue-same-route",
-                        "reroute-cheaper",
-                        "reroute-stronger",
-                        "split-task",
-                        "abort",
-                    ]
-                },
-            },
-            "forbidDowngradeForCriticalReview": {"type": "boolean"},
-            "maxAutoReroutesPerTask": {"type": "integer", "minimum": 0},
-            "defaultAutoAction": {"type": "string"},
-            "budgetModes": {"type": "object"},
-        },
-    ),
-    "agent-lifecycle-budget-decision-receipt.v1": _open_object_schema(
-        "agent-lifecycle-budget-decision-receipt.v1",
-        required=[
-            "schemaVersion",
-            "runId",
-            "taskId",
-            "attemptId",
-            "sourceRevision",
-            "priorRouteDecisionDigest",
-            "usageReceiptDigest",
-            "overrunReason",
-            "decisionMode",
-            "selectedAction",
-            "resumePhase",
-            "expectedWorkflowRevision",
-        ],
-        properties={
-            "runId": {"type": "string", "minLength": 1},
-            "taskId": {"type": "string", "minLength": 1},
-            "attemptId": {"type": "string", "minLength": 1},
-            "sourceRevision": {"type": "string", "minLength": 1},
-            "priorRouteDecisionDigest": {"type": "string", "minLength": 64, "maxLength": 64},
-            "usageReceiptDigest": {"type": "string", "minLength": 64, "maxLength": 64},
-            "overrunReason": {"type": "string", "minLength": 1},
-            "decisionMode": {"enum": ["manual", "auto"]},
-            "selectedAction": {"type": "string", "minLength": 1},
-            "operatorIdentityHash": {"type": "string"},
-            "policyDigest": {"type": "string"},
-            "nextRouteDecisionDigest": {"type": ["string", "null"]},
-            "capDeltas": {"type": "object"},
-            "resumePhase": {"type": "string", "minLength": 1},
-            "expectedWorkflowRevision": {"type": "integer", "minimum": 0},
-        },
-    ),
-    "agent-cursor-compat-evidence.v1": _open_object_schema(
-        "agent-cursor-compat-evidence.v1",
-        required=["schemaVersion", "status", "source", "checks", "productionPromotionClaimed"],
-        properties={
-            "status": {"enum": ["PASS", "FAIL"]},
-            "source": {"type": "object"},
-            "checks": {"type": "array", "items": {"type": "object"}},
-            "providerModelHashes": {"type": "object"},
-            "productionPromotionClaimed": {"const": False},
-        },
-    ),
-    "agent-lifecycle-model-route-request.v1": {
-        "$schema": "https://json-schema.org/draft/2020-12/schema",
-        "$id": "agent-lifecycle-model-route-request.v1",
-        "type": "object",
-        "additionalProperties": True,
-        "required": ["schemaVersion", "operationId", "phase", "sddTier"],
-        "properties": {
-            "schemaVersion": {"const": "agent-lifecycle-model-route-request.v1"},
-            "operationId": {"type": "string", "minLength": 1},
-            "phase": {"type": "string", "minLength": 1},
-            "sddTier": {"enum": ["S0", "S1", "S2"]},
-        },
-    },
-    "agent-lifecycle-model-route-decision.v1": {
-        "$schema": "https://json-schema.org/draft/2020-12/schema",
-        "$id": "agent-lifecycle-model-route-decision.v1",
-        "type": "object",
-        "additionalProperties": True,
-        "required": ["schemaVersion", "operationId", "modelClass", "decisionDigest"],
-        "properties": {
-            "schemaVersion": {"const": "agent-lifecycle-model-route-decision.v1"},
-            "operationId": {"type": "string", "minLength": 1},
-            "modelClass": {"type": "string", "minLength": 1},
-            "decisionDigest": {"type": "string", "minLength": 64, "maxLength": 64},
-        },
-    },
-    "agent-lifecycle-model-usage-receipt.v1": {
-        "$schema": "https://json-schema.org/draft/2020-12/schema",
-        "$id": "agent-lifecycle-model-usage-receipt.v1",
-        "type": "object",
-        "additionalProperties": True,
-        "required": ["schemaVersion", "operationId", "host", "modelClass", "providerModelHash", "usage", "attestation"],
-        "properties": {
-            "schemaVersion": {"const": "agent-lifecycle-model-usage-receipt.v1"},
-            "operationId": {"type": "string", "minLength": 1},
-            "host": {"type": "string", "minLength": 1},
-            "modelClass": {"type": "string", "minLength": 1},
-            "providerModelHash": {"type": "string", "minLength": 1},
-            "usage": {"type": "object"},
-            "attestation": {"type": "object"},
-        },
-    },
-    "agent-release-candidate-inventory.v1": _open_object_schema(
-        "agent-release-candidate-inventory.v1",
-        required=[
-            "schemaVersion",
-            "packageId",
-            "planRevision",
-            "planDigest",
-            "files",
-            "candidatePayloadInventoryDigest",
-        ],
-        properties={
-            "packageId": {"type": "string", "minLength": 1},
-            "planRevision": {"type": "integer", "minimum": 1},
-            "planDigest": {"type": "string", "minLength": 64, "maxLength": 64},
-            "files": {"type": "array", "items": {"type": "object"}},
-            "candidatePayloadInventoryDigest": {"type": "string", "minLength": 64, "maxLength": 64},
-        },
-    ),
-    "agent-release-assembly-evidence.v1": _open_object_schema(
-        "agent-release-assembly-evidence.v1",
-        required=["schemaVersion", "status", "inventory", "productionPromotionClaimed"],
-        properties={
-            "status": {"enum": ["PASS", "FAIL"]},
-            "inventory": {"type": "object"},
-            "productionPromotionClaimed": {"const": False},
-        },
-    ),
-    "agent-release-verification-evidence.v1": _open_object_schema(
-        "agent-release-verification-evidence.v1",
-        required=["schemaVersion", "status", "inventory", "mismatches", "productionPromotionClaimed"],
-        properties={
-            "status": {"enum": ["PASS", "FAIL"]},
-            "inventory": {"type": "object"},
-            "mismatches": {"type": "array", "items": {"type": "object"}},
-            "productionPromotionClaimed": {"const": False},
-        },
-    ),
-    "agent-final-candidate-audit.v1": _open_object_schema(
-        "agent-final-candidate-audit.v1",
-        required=["schemaVersion", "status", "planRevision", "planDigest", "productionPromotionClaimed"],
-        properties={
-            "status": {"enum": ["PASS", "FAIL"]},
-            "semanticStatus": {"type": "string", "minLength": 1},
-            "planRevision": {"type": "integer", "minimum": 1},
-            "planDigest": {"type": "string", "minLength": 64, "maxLength": 64},
-            "productionPromotionClaimed": {"const": False},
-        },
-    ),
-    "agent-support-matrix-contract-evidence.v1": _open_object_schema(
-        "agent-support-matrix-contract-evidence.v1",
-        required=["schemaVersion", "status", "adapterMaturity", "productionPromotionClaimed"],
-        properties={
-            "status": {"enum": ["PASS", "FAIL"]},
-            "adapterMaturity": {"type": "string", "minLength": 1},
-            "productionPromotionClaimed": {"const": False},
-        },
-    ),
-    "agent-deferred-promotion-contract-evidence.v1": _open_object_schema(
-        "agent-deferred-promotion-contract-evidence.v1",
-        required=["schemaVersion", "status", "deferredProductionPromotion", "productionPromotionClaimed"],
-        properties={
-            "status": {"enum": ["PASS", "FAIL"]},
-            "deferredProductionPromotion": {"const": True},
-            "productionPromotionClaimed": {"const": False},
-        },
-    ),
-    "agent-neutrality-report.v1": _open_object_schema(
-        "agent-neutrality-report.v1",
-        required=["schemaVersion", "counters"],
-        properties={
-            "counters": {"type": "object"},
-            "operation": {"type": "object"},
-        },
-    ),
-    "agent-live-calibration-verification.v1": _open_object_schema(
-        "agent-live-calibration-verification.v1",
-        required=["schemaVersion", "status", "blockers", "productionPromotionClaimed"],
-        properties={
-            "status": {"enum": ["PASS", "FAIL"]},
-            "blockers": {"type": "array", "items": {"type": "object"}},
-            "productionPromotionClaimed": {"const": False},
-        },
-    ),
-    "agent-lifecycle-live-calibration-receipt.v1": _open_object_schema(
-        "agent-lifecycle-live-calibration-receipt.v1",
-        required=["schemaVersion", "status", "host", "runs", "syntheticReplayUsed"],
-        properties={
-            "status": {"enum": ["PASS", "FAIL"]},
-            "host": {"type": "string", "minLength": 1},
-            "runs": {"type": "array", "items": {"type": "object"}},
-            "syntheticReplayUsed": {"const": False},
-        },
-    ),
-    "agent-lifecycle-live-host-conformance-receipt.v1": _open_object_schema(
-        "agent-lifecycle-live-host-conformance-receipt.v1",
-        required=["schemaVersion", "status", "host", "operations", "syntheticReplayUsed", "usageAttested"],
-        properties={
-            "status": {"enum": ["PASS", "FAIL"]},
-            "host": {"type": "string", "minLength": 1},
-            "operations": {"type": "array", "items": {"type": "object"}},
-            "syntheticReplayUsed": {"const": False},
-            "usageAttested": {"const": True},
-        },
-    ),
-    "agent-live-host-conformance-verification.v1": _open_object_schema(
-        "agent-live-host-conformance-verification.v1",
-        required=["schemaVersion", "status", "promotedHosts", "checks", "blockers", "productionPromotionClaimed"],
-        properties={
-            "status": {"enum": ["PASS", "FAIL"]},
-            "promotedHosts": {"type": "array", "items": {"type": "string"}},
-            "checks": {"type": "array", "items": {"type": "object"}},
-            "blockers": {"type": "array", "items": {"type": "object"}},
-            "productionPromotionClaimed": {"const": False},
-        },
-    ),
-    "agent-live-host-promotion-plan.v1": _open_object_schema(
-        "agent-live-host-promotion-plan.v1",
-        required=[
-            "schemaVersion",
-            "packageId",
-            "sddTier",
-            "status",
-            "intent",
-            "hostOrder",
-            "sequencingPolicy",
-            "hostAvailabilitySnapshot",
-            "sharedInputs",
-            "artifactRootPolicy",
-            "budgetPolicy",
-            "blockerCodes",
-            "operationEvidenceRequirements",
-            "validationCommands",
-            "evidenceArtifacts",
-            "workstreams",
-            "acceptanceCriteria",
-        ],
-        properties={
-            "packageId": {"type": "string", "minLength": 1},
-            "sddTier": {"enum": ["S0", "S1", "S2"]},
-            "status": {"enum": ["DRAFT", "READY", "FROZEN", "BLOCKED"]},
-            "intent": {"type": "string", "minLength": 1},
-            "hostOrder": {"type": "array", "items": {"type": "string", "minLength": 1}},
-            "sequencingPolicy": {"type": "object"},
-            "hostAvailabilitySnapshot": {"type": "object"},
-            "sharedInputs": {"type": "object"},
-            "artifactRootPolicy": {"type": "object"},
-            "budgetPolicy": {"type": "object"},
-            "blockerCodes": {"type": "array", "items": {"type": "string", "minLength": 1}},
-            "operationEvidenceRequirements": {"type": "object"},
-            "validationCommands": {"type": "array", "items": {"type": "object"}},
-            "evidenceArtifacts": {"type": "array", "items": {"type": "object"}},
-            "workstreams": {"type": "array", "items": {"type": "object"}},
-            "acceptanceCriteria": {"type": "array", "items": {"type": "object"}},
-        },
-    ),
-    "agent-live-host-promotion-plan-validation.v1": _open_object_schema(
-        "agent-live-host-promotion-plan-validation.v1",
-        required=["schemaVersion", "status", "plan", "checks", "blockers", "productionPromotionClaimed"],
-        properties={
-            "status": {"enum": ["PASS", "FAIL"]},
-            "plan": {"type": "object"},
-            "checks": {"type": "array", "items": {"type": "object"}},
-            "blockers": {"type": "array", "items": {"type": "object"}},
-            "productionPromotionClaimed": {"const": False},
-        },
-    ),
-    "agent-digest-authority-evidence.v1": _open_object_schema(
-        "agent-digest-authority-evidence.v1",
-        required=["schemaVersion", "status", "canonicalAuthority", "checks", "productionPromotionClaimed"],
-        properties={
-            "status": {"enum": ["PASS", "FAIL"]},
-            "canonicalAuthority": {"type": "string", "minLength": 1},
-            "checks": {"type": "array", "items": {"type": "object"}},
-            "productionPromotionClaimed": {"const": False},
-        },
-    ),
-    "agent-docs-compat-evidence.v1": _open_object_schema(
-        "agent-docs-compat-evidence.v1",
-        required=["schemaVersion", "status", "checks", "blockers", "productionPromotionClaimed"],
-        properties={
-            "status": {"enum": ["PASS", "FAIL"]},
-            "checks": {"type": "array", "items": {"type": "object"}},
-            "blockers": {"type": "array", "items": {"type": "object"}},
-            "productionPromotionClaimed": {"const": False},
-        },
-    ),
-    "agent-negative-suite-coverage.v1": _open_object_schema(
-        "agent-negative-suite-coverage.v1",
-        required=["schemaVersion", "status", "expectedRange", "coveredScenarios"],
-        properties={
-            "status": {"enum": ["PASS", "FAIL"]},
-            "expectedRange": {"type": "string", "minLength": 1},
-            "coveredScenarios": {"type": "array", "items": {"type": "object"}},
-        },
-    ),
-    "agent-task-packet-context-fit.v1": _open_object_schema(
-        "agent-task-packet-context-fit.v1",
-        required=["schemaVersion", "status", "targetWindows", "checks"],
-        properties={
-            "status": {"enum": ["PASS", "FAIL"]},
-            "targetWindows": {"type": "array", "items": {"type": "string"}},
-            "checks": {"type": "array", "items": {"type": "object"}},
-        },
-    ),
-    "agent-packaging-smoke-evidence.v1": _open_object_schema(
-        "agent-packaging-smoke-evidence.v1",
-        required=["schemaVersion", "status", "commands", "blockers", "productionPromotionClaimed"],
-        properties={
-            "status": {"enum": ["PASS", "FAIL"]},
-            "commands": {"type": "array", "items": {"type": "object"}},
-            "blockers": {"type": "array", "items": {"type": "object"}},
-            "productionPromotionClaimed": {"const": False},
-        },
-    ),
-    "agent-adapter-scaffold-result.v1": _open_object_schema(
-        "agent-adapter-scaffold-result.v1",
-        required=["schemaVersion", "status", "host", "maturity", "files", "productionPromotionClaimed"],
-        properties={
-            "status": {"enum": ["PASS", "DRY_RUN"]},
-            "host": {"type": "string", "minLength": 1},
-            "maturity": {"const": "EXPERIMENTAL"},
-            "files": {"type": "array", "items": {"type": "object"}},
-            "productionPromotionClaimed": {"const": False},
-        },
-    ),
-    "agent-workflow-lineage-check.v1": _open_object_schema(
-        "agent-workflow-lineage-check.v1",
-        required=["schemaVersion", "status", "planRevision", "planDigest", "lineageChecks"],
-        properties={
-            "status": {"enum": ["PASS", "FAIL"]},
-            "planRevision": {"type": "integer", "minimum": 1},
-            "planDigest": {"type": "string", "minLength": 64, "maxLength": 64},
-            "lineageChecks": {"type": "array", "items": {"type": "object"}},
-        },
-    ),
-    "agent-public-contract-policy.v1": _open_object_schema(
-        "agent-public-contract-policy.v1",
-        required=[
-            "schemaVersion",
-            "status",
-            "rules",
-            "requiredCoreSchemas",
-            "schemas",
-            "cliOutputs",
-            "productionPromotionClaimed",
-            "policyDigest",
-        ],
-        properties={
-            "status": {"const": "PASS"},
-            "rules": {"type": "object"},
-            "requiredCoreSchemas": {"type": "array", "items": {"type": "string", "minLength": 1}},
-            "schemas": {"type": "array", "items": {"type": "object"}},
-            "cliOutputs": {"type": "array", "items": {"type": "object"}},
-            "productionPromotionClaimed": {"const": False},
-            "policyDigest": {"type": "string", "minLength": 64, "maxLength": 64},
-        },
-    ),
-    "agent-public-contract-policy-validation.v1": _open_object_schema(
-        "agent-public-contract-policy-validation.v1",
-        required=[
-            "schemaVersion",
-            "status",
-            "schemaCount",
-            "cliOutputCount",
-            "deprecatedCompatibleSchemas",
-            "blockers",
-            "policyDigest",
-            "validationDigest",
-        ],
-        properties={
-            "status": {"enum": ["PASS", "FAIL"]},
-            "schemaCount": {"type": "integer", "minimum": 0},
-            "cliOutputCount": {"type": "integer", "minimum": 0},
-            "deprecatedCompatibleSchemas": {"type": "array", "items": {"type": "string"}},
-            "blockers": {"type": "array", "items": {"type": "object"}},
-            "policyDigest": {"type": ["string", "null"]},
-            "validationDigest": {"type": "string", "minLength": 64, "maxLength": 64},
-        },
-    ),
-    "agent-lifecycle-cost-report.v1": _open_object_schema(
-        "agent-lifecycle-cost-report.v1",
-        required=["schemaVersion", "mode", "entries"],
-        properties={
-            "mode": {"enum": ["light", "standard", "strict", "release"]},
-            "entries": {"type": "array", "items": {"type": "object"}},
-            "limits": {"type": "object"},
-            "overLimitReason": {"type": "string"},
-            "productionPromotionClaimed": {"const": False},
-        },
-    ),
-    "agent-lifecycle-cost-validation.v1": _open_object_schema(
-        "agent-lifecycle-cost-validation.v1",
-        required=["schemaVersion", "status", "mode", "totals", "ratios", "limits", "blockers", "reportDigest"],
-        properties={
-            "status": {"enum": ["PASS", "FAIL"]},
-            "mode": {"type": ["string", "null"]},
-            "totals": {"type": "object"},
-            "ratios": {"type": "object"},
-            "limits": {"type": "object"},
-            "blockers": {"type": "array", "items": {"type": "object"}},
-            "reportDigest": {"type": "string", "minLength": 64, "maxLength": 64},
-        },
-    ),
-    "agent-plan-reference-validation.v1": _open_object_schema(
-        "agent-plan-reference-validation.v1",
-        required=["schemaVersion", "status", "packageId", "referenceCount", "repositoryIds", "blockers", "validationDigest"],
-        properties={
-            "status": {"enum": ["PASS", "FAIL"]},
-            "packageId": {"type": ["string", "null"]},
-            "referenceCount": {"type": "integer", "minimum": 0},
-            "repositoryIds": {"type": "array", "items": {"type": "string", "minLength": 1}},
-            "blockers": {"type": "array", "items": {"type": "object"}},
-            "validationDigest": {"type": "string", "minLength": 64, "maxLength": 64},
-        },
-    ),
-    "agent-plan-snapshot.v1": _open_object_schema(
-        "agent-plan-snapshot.v1",
-        required=[
-            "schemaVersion",
-            "status",
-            "packageId",
-            "planRevision",
-            "sourceDigest",
-            "baseRevision",
-            "specificationDigest",
-            "acceptanceDigest",
-            "repositoryReferencesDigest",
-            "referenceValidationDigest",
-            "immutable",
-            "snapshotDigest",
-        ],
-        properties={
-            "status": {"const": "PASS"},
-            "packageId": {"type": ["string", "null"]},
-            "planRevision": {"type": ["integer", "null"], "minimum": 0},
-            "sourceDigest": {"type": "string", "minLength": 64, "maxLength": 64},
-            "baseRevision": {"type": "object"},
-            "specificationDigest": {"type": "string", "minLength": 64, "maxLength": 64},
-            "acceptanceDigest": {"type": "string", "minLength": 64, "maxLength": 64},
-            "repositoryReferencesDigest": {"type": "string", "minLength": 64, "maxLength": 64},
-            "referenceValidationDigest": {"type": "string", "minLength": 64, "maxLength": 64},
-            "immutable": {"const": True},
-            "snapshotDigest": {"type": "string", "minLength": 64, "maxLength": 64},
-        },
-    ),
-    "agent-plan-reconciliation.v1": _open_object_schema(
-        "agent-plan-reconciliation.v1",
-        required=["schemaVersion", "status", "classification", "packageId", "snapshotDigest", "currentDigest", "blockers", "reconciliationDigest"],
-        properties={
-            "status": {"enum": ["PASS", "FAIL"]},
-            "classification": {"enum": ["MATCH", "REQUIRES_NEW_PLAN", "BLOCKED"]},
-            "packageId": {"type": ["string", "null"]},
-            "snapshotDigest": {"type": ["string", "null"]},
-            "currentDigest": {"type": "string", "minLength": 64, "maxLength": 64},
-            "blockers": {"type": "array", "items": {"type": "object"}},
-            "reconciliationDigest": {"type": "string", "minLength": 64, "maxLength": 64},
-        },
-    ),
-    "agent-plan-handoff.v1": _open_object_schema(
-        "agent-plan-handoff.v1",
-        required=[
-            "schemaVersion",
-            "status",
-            "packageId",
-            "planRevision",
-            "planStatus",
-            "repositoryReferences",
-            "referenceValidation",
-            "workstreams",
-            "acceptanceIds",
-            "omitted",
-            "sourceDigest",
-            "estimatedTokens",
-            "targetTokens",
-            "handoffDigest",
-        ],
-        properties={
-            "status": {"enum": ["PASS", "FAIL"]},
-            "packageId": {"type": ["string", "null"]},
-            "planRevision": {"type": ["integer", "null"], "minimum": 0},
-            "planStatus": {"type": ["string", "null"]},
-            "snapshotDigest": {"type": ["string", "null"]},
-            "repositoryReferences": {"type": "array", "items": {"type": "object"}},
-            "referenceValidation": {"type": "object"},
-            "workstreams": {"type": "array", "items": {"type": "object"}},
-            "acceptanceIds": {"type": "array", "items": {"type": "string", "minLength": 1}},
-            "omitted": {"type": "object"},
-            "sourceDigest": {"type": "string", "minLength": 64, "maxLength": 64},
-            "estimatedTokens": {"type": "integer", "minimum": 1},
-            "targetTokens": {"type": "integer", "minimum": 1},
-            "handoffDigest": {"type": "string", "minLength": 64, "maxLength": 64},
-        },
-    ),
-}
+_SCHEMAS: dict[str, dict[str, Any]] = {}
+for _group in _SCHEMA_GROUPS:
+    _overlap = set(_SCHEMAS).intersection(_group)
+    if _overlap:
+        raise RuntimeError(f"duplicate schema ids: {sorted(_overlap)}")
+    _SCHEMAS.update(_group)
 
 
 def list_schemas() -> dict[str, Any]:
