+++ src/agent_lifecycle/contracts/status_goal_schemas.py
@@ -1,0 +1,259 @@
+"""Built-in JSON schema definitions for a lifecycle contract domain."""
+
+from __future__ import annotations
+
+from typing import Any
+
+from agent_lifecycle.contracts.schema_builders import open_object_schema as _open_object_schema
+
+STATUS_GOAL_SCHEMAS: dict[str, dict[str, Any]] = {
+    "agent-diagnostic-bundle.v1": _open_object_schema(
+        "agent-diagnostic-bundle.v1",
+        required=[
+            "schemaVersion",
+            "status",
+            "projectRoot",
+            "sourceOfTruth",
+            "redacted",
+            "artifacts",
+            "blockers",
+            "productionPromotionClaimed",
+        ],
+        properties={
+            "status": {"enum": ["PASS", "FAIL"]},
+            "projectRoot": {"const": "<checkout>"},
+            "sourceOfTruth": {"const": False},
+            "redacted": {"const": True},
+            "artifacts": {"type": "array", "items": {"type": "object"}},
+            "blockers": {"type": "array", "items": {"type": "object"}},
+            "productionPromotionClaimed": {"const": False},
+        },
+    ),
+    "agent-readonly-status-view.v1": _open_object_schema(
+        "agent-readonly-status-view.v1",
+        required=[
+            "schemaVersion",
+            "status",
+            "sourceOfTruth",
+            "projectRoot",
+            "itemCount",
+            "items",
+            "blockers",
+            "productionPromotionClaimed",
+        ],
+        properties={
+            "status": {"enum": ["PASS", "FAIL"]},
+            "sourceOfTruth": {"const": False},
+            "projectRoot": {"const": "<checkout>"},
+            "itemCount": {"type": "integer", "minimum": 0},
+            "items": {"type": "array", "items": {"type": "object"}},
+            "blockers": {"type": "array", "items": {"type": "object"}},
+            "productionPromotionClaimed": {"const": False},
+        },
+    ),
+    "agent-completion-signal.v1": {
+        "$schema": "https://json-schema.org/draft/2020-12/schema",
+        "$id": "agent-completion-signal.v1",
+        "type": "object",
+        "additionalProperties": True,
+        "required": [
+            "schemaVersion",
+            "runId",
+            "packageId",
+            "planRevision",
+            "planDigest",
+            "sourceRevision",
+            "status",
+            "evidenceIds",
+            "verifier",
+        ],
+        "properties": {
+            "schemaVersion": {"const": "agent-completion-signal.v1"},
+            "runId": {"type": "string", "minLength": 1},
+            "packageId": {"type": "string", "minLength": 1},
+            "planRevision": {"type": "integer", "minimum": 1},
+            "planDigest": {"type": "string", "minLength": 64, "maxLength": 64},
+            "sourceRevision": {"type": "string", "minLength": 1},
+            "status": {"enum": ["PASS", "WAIVED", "FAIL"]},
+            "evidenceIds": {"type": "array", "items": {"type": "string", "minLength": 1}},
+            "verifier": {"type": "object"},
+            "waiver": {"type": "object"},
+        },
+    },
+    "agent-completion-signal-validation.v1": _open_object_schema(
+        "agent-completion-signal-validation.v1",
+        required=["schemaVersion", "status", "signalStatus", "signalDigest", "waived"],
+        properties={
+            "status": {"const": "PASS"},
+            "signalStatus": {"enum": ["PASS", "WAIVED"]},
+            "signalDigest": {"type": "string", "minLength": 64, "maxLength": 64},
+            "waived": {"type": "boolean"},
+        },
+    ),
+    "agent-completion-check.v1": _open_object_schema(
+        "agent-completion-check.v1",
+        required=["schemaVersion", "checkId", "kind", "description", "receiptPath", "requiredEvidenceIds"],
+        properties={
+            "checkId": {"type": "string", "minLength": 1},
+            "kind": {"enum": ["verification", "external-action"]},
+            "description": {"type": "string", "minLength": 1},
+            "receiptPath": {"type": "string", "minLength": 1},
+            "requiredEvidenceIds": {"type": "array", "items": {"type": "string", "minLength": 1}},
+        },
+    ),
+    "agent-completion-check-validation.v1": _open_object_schema(
+        "agent-completion-check-validation.v1",
+        required=["schemaVersion", "checkId", "kind", "receiptPath", "requiredEvidenceIds", "checkDigest"],
+        properties={
+            "checkId": {"type": "string", "minLength": 1},
+            "kind": {"enum": ["verification", "external-action"]},
+            "receiptPath": {"type": "string", "minLength": 1},
+            "requiredEvidenceIds": {"type": "array", "items": {"type": "string", "minLength": 1}},
+            "checkDigest": {"type": "string", "minLength": 64, "maxLength": 64},
+        },
+    ),
+    "agent-completion-check-receipt.v1": _open_object_schema(
+        "agent-completion-check-receipt.v1",
+        required=[
+            "schemaVersion",
+            "checkId",
+            "runId",
+            "packageId",
+            "planRevision",
+            "planDigest",
+            "sourceRevision",
+            "status",
+            "evidenceIds",
+            "verifier",
+        ],
+        properties={
+            "checkId": {"type": "string", "minLength": 1},
+            "runId": {"type": "string", "minLength": 1},
+            "packageId": {"type": "string", "minLength": 1},
+            "planRevision": {"type": "integer", "minimum": 1},
+            "planDigest": {"type": "string", "minLength": 64, "maxLength": 64},
+            "sourceRevision": {"type": "string", "minLength": 1},
+            "status": {"enum": ["PASS", "FAIL"]},
+            "evidenceIds": {"type": "array", "items": {"type": "string", "minLength": 1}},
+            "verifier": {"type": "object"},
+            "externalActionReceipt": {"type": "object"},
+        },
+    ),
+    "agent-completion-check-receipt-validation.v1": _open_object_schema(
+        "agent-completion-check-receipt-validation.v1",
+        required=["schemaVersion", "status", "checkId", "checkKind", "evidenceIds", "receiptDigest", "checkDigest"],
+        properties={
+            "status": {"const": "PASS"},
+            "checkId": {"type": "string", "minLength": 1},
+            "checkKind": {"enum": ["verification", "external-action"]},
+            "evidenceIds": {"type": "array", "items": {"type": "string", "minLength": 1}},
+            "receiptDigest": {"type": "string", "minLength": 64, "maxLength": 64},
+            "checkDigest": {"type": "string", "minLength": 64, "maxLength": 64},
+        },
+    ),
+    "agent-goal-record.v1": _open_object_schema(
+        "agent-goal-record.v1",
+        required=[
+            "schemaVersion",
+            "goalId",
+            "userIntent",
+            "ownerOutcome",
+            "constraints",
+            "status",
+            "lineage",
+            "evidenceIds",
+            "updatedAt",
+        ],
+        properties={
+            "goalId": {"type": "string", "minLength": 1},
+            "userIntent": {"type": "string", "minLength": 1},
+            "ownerOutcome": {"type": "string", "minLength": 1},
+            "constraints": {"type": "array", "items": {"type": "string", "minLength": 1}},
+            "status": {"enum": ["ACTIVE", "BLOCKED", "READY_FOR_FINALIZATION", "COMPLETE"]},
+            "lineage": {"type": "object"},
+            "completionCheck": {"type": "object"},
+            "evidenceIds": {"type": "array", "items": {"type": "string", "minLength": 1}},
+            "updatedAt": {"type": "string", "minLength": 1},
+        },
+    ),
+    "agent-goal-record-validation.v1": _open_object_schema(
+        "agent-goal-record-validation.v1",
+        required=["schemaVersion", "status", "goalId", "goalStatus", "lineage", "stateRevision", "goalDigest"],
+        properties={
+            "status": {"const": "PASS"},
+            "goalId": {"type": "string", "minLength": 1},
+            "goalStatus": {"enum": ["ACTIVE", "BLOCKED", "READY_FOR_FINALIZATION", "COMPLETE"]},
+            "lineage": {"type": "object"},
+            "stateRevision": {"type": "integer", "minimum": 1},
+            "goalDigest": {"type": "string", "minLength": 64, "maxLength": 64},
+        },
+    ),
+    "agent-objective-snapshot.v1": _open_object_schema(
+        "agent-objective-snapshot.v1",
+        required=["schemaVersion", "status", "goal", "lifecycle", "estimatedTokens", "snapshotDigest"],
+        properties={
+            "status": {"const": "PASS"},
+            "goal": {"type": "object"},
+            "lifecycle": {"type": "object"},
+            "completionCheck": {"type": ["object", "null"]},
+            "estimatedTokens": {"type": "integer", "minimum": 1},
+            "snapshotDigest": {"type": "string", "minLength": 64, "maxLength": 64},
+        },
+    ),
+    "agent-follow-up-register.v1": _open_object_schema(
+        "agent-follow-up-register.v1",
+        required=["schemaVersion", "lineage", "items"],
+        properties={
+            "lineage": {"type": "object"},
+            "items": {"type": "array", "items": {"type": "object"}},
+            "updatedAt": {"type": "string", "minLength": 1},
+        },
+    ),
+    "agent-follow-up-register-validation.v1": _open_object_schema(
+        "agent-follow-up-register-validation.v1",
+        required=[
+            "schemaVersion",
+            "status",
+            "registerDigest",
+            "lineage",
+            "itemCount",
+            "openItemIds",
+            "finalizationBlockers",
+            "items",
+            "validationDigest",
+        ],
+        properties={
+            "status": {"const": "PASS"},
+            "registerDigest": {"type": "string", "minLength": 64, "maxLength": 64},
+            "lineage": {"type": "object"},
+            "itemCount": {"type": "integer", "minimum": 0},
+            "openItemIds": {"type": "array", "items": {"type": "string", "minLength": 1}},
+            "finalizationBlockers": {"type": "array", "items": {"type": "object"}},
+            "items": {"type": "array", "items": {"type": "object"}},
+            "validationDigest": {"type": "string", "minLength": 64, "maxLength": 64},
+        },
+    ),
+    "agent-follow-up-close-result.v1": _open_object_schema(
+        "agent-follow-up-close-result.v1",
+        required=["schemaVersion", "status", "itemId", "registerDigest"],
+        properties={
+            "status": {"const": "PASS"},
+            "itemId": {"type": "string", "minLength": 1},
+            "registerDigest": {"type": "string", "minLength": 64, "maxLength": 64},
+        },
+    ),
+    "agent-follow-up-summary.v1": _open_object_schema(
+        "agent-follow-up-summary.v1",
+        required=["schemaVersion", "status", "lineage", "registerDigest", "counts", "openItems", "finalizationBlockers", "estimatedTokens", "summaryDigest"],
+        properties={
+            "status": {"const": "PASS"},
+            "lineage": {"type": "object"},
+            "registerDigest": {"type": "string", "minLength": 64, "maxLength": 64},
+            "counts": {"type": "object"},
+            "openItems": {"type": "array", "items": {"type": "object"}},
+            "finalizationBlockers": {"type": "array", "items": {"type": "object"}},
+            "estimatedTokens": {"type": "integer", "minimum": 1},
+            "summaryDigest": {"type": "string", "minLength": 64, "maxLength": 64},
+        },
+    ),
+}
