+++ src/agent_lifecycle/metrics/__init__.py
@@ -6,10 +6,13 @@ from agent_lifecycle.metrics.costs import (
     require_lifecycle_cost_pass,
     validate_lifecycle_cost_report,
 )
+from agent_lifecycle.metrics.cost_collection import build_lifecycle_cost_summary, generate_lifecycle_cost_report
 
 __all__ = [
     "COST_CATEGORIES",
     "DEFAULT_MODE_LIMITS",
+    "build_lifecycle_cost_summary",
+    "generate_lifecycle_cost_report",
     "require_lifecycle_cost_pass",
     "validate_lifecycle_cost_report",
 ]
