+++ tests/cli/test_contract_metrics_commands.py
@@ -64,6 +64,55 @@ class CliContractMetricsCommandTests(unittest.TestCase):
         self.assertEqual(payload["status"], "PASS")
         self.assertLessEqual(payload["ratios"]["pipelineTokenShare"], payload["limits"]["maxPipelineTokenShare"])
 
+    def test_metrics_cost_report_cli_writes_report_and_compact_summary(self) -> None:
+        with tempfile.TemporaryDirectory() as tmp:
+            root = Path(tmp)
+            artifact = root / "usage.json"
+            report_path = root / "generated-cost.json"
+            summary_path = root / "cost-summary.json"
+            task_packet = root / "task-packet.json"
+            artifact.write_text(json.dumps(_usage_receipt()), encoding="utf-8")
+            task_packet.write_text(json.dumps(_cost_summary_task_packet()), encoding="utf-8")
+
+            code, payload = _run_cli(
+                [
+                    "metrics",
+                    "cost-report",
+                    "--project-root",
+                    str(root),
+                    "--artifact",
+                    str(artifact),
+                    "--out",
+                    str(report_path),
+                    "--summary-out",
+                    str(summary_path),
+                ]
+            )
+            context_code, context_payload = _run_cli(
+                [
+                    "context",
+                    "check",
+                    "--profile",
+                    str(ROOT / "profiles/small-context-profile.v1.json"),
+                    "--task-packet",
+                    str(task_packet),
+                    "--summary",
+                    str(summary_path),
+                    "--target-window",
+                    "4k-strict",
+                ]
+            )
+            self.assertTrue(report_path.is_file())
+            self.assertTrue(summary_path.is_file())
+
+        self.assertEqual(code, 0)
+        self.assertEqual(payload["schemaVersion"], "agent-lifecycle-cost-generation.v1")
+        self.assertEqual(payload["status"], "PASS")
+        self.assertFalse(payload["liveCallsStarted"])
+        self.assertFalse(payload["productionPromotionClaimed"])
+        self.assertEqual(context_code, 0)
+        self.assertEqual(context_payload["status"], "PASS")
+
 
 def _cost_report() -> dict[str, object]:
     return {
@@ -76,6 +125,67 @@ def _cost_report() -> dict[str, object]:
             {"category": "coordination", "tokens": 600, "steps": 1},
         ],
         "productionPromotionClaimed": False,
+    }
+
+
+def _usage_receipt() -> dict[str, object]:
+    return {
+        "schemaVersion": "agent-lifecycle-model-usage-receipt.v1",
+        "operationId": "impl-op",
+        "runId": "run",
+        "packageId": "package",
+        "taskId": "WS-01",
+        "planDigest": "0" * 64,
+        "sourceRevision": "source",
+        "host": "codex",
+        "modelClass": "standard-code",
+        "providerModelHash": "redacted",
+        "usage": {
+            "inputTokens": 100,
+            "outputTokens": 20,
+            "billableTokens": 120,
+            "cumulativeContextBytes": 4096,
+            "toolCalls": 1,
+            "wallSeconds": 2,
+        },
+        "attestation": {"source": "host", "status": "ATTESTED"},
+    }
+
+
+def _cost_summary_task_packet() -> dict[str, object]:
+    return {
+        "schemaVersion": "agent-task-packet.v1",
+        "plan": {"packageId": "package", "planRevision": 1, "planDigest": "0" * 64},
+        "task": {
+            "id": "WS-01",
+            "title": "Review lifecycle cost",
+            "owner": "worker",
+            "reviewer": "reviewer",
+            "dependsOn": [],
+            "required": True,
+            "plannedItems": ["R-COST"],
+            "acceptanceIds": ["AC-COST"],
+            "evidenceIds": ["EV-COST"],
+            "artifactPaths": {},
+            "capabilityHints": [],
+            "requiredTools": [],
+            "executionPolicy": {},
+        },
+        "ownership": {
+            "writes": ["src/agent_lifecycle/metrics"],
+            "readOnly": ["profiles/small-context-profile.v1.json"],
+            "forbiddenWrites": [],
+            "leadOwned": [],
+        },
+        "specification": {
+            "tier": "S1",
+            "revision": 1,
+            "requirements": ["R-COST"],
+            "traceDigest": "1" * 64,
+        },
+        "context": {"refs": ["profiles/small-context-profile.v1.json"]},
+        "validation": {"acceptanceIds": ["AC-COST"], "evidenceIds": ["EV-COST"]},
+        "acceptance": [{"id": "AC-COST", "statement": "cost summary fits"}],
     }
